refactor: split the ferriki-core monolith into focused modules
lib.rs shrinks from 5,174 to 808 lines and keeps only the crate root
and the N-API facade (FerrikiHighlighter, create_highlighter,
ferriki_version) plus the existing test module. The code moves verbatim
along the pre-existing section banners into:

- types.rs (shared data types, RuleId architecture)
- grammar.rs (grammar parsing and compilation entry)
- theme.rs (scope resolution, ThemeCache)
- injection.rs (injection selector parsing and compilation)
- rule.rs (rule compilation)
- scanner.rs (scanner compilation and ordered matching)
- tokenize.rs (tokenization loop and helpers)
- render.rs (UTF-16 mapping, HTML/JSON/HAST rendering)

Visibility is adjusted minimally (moved items and their fields become
pub(crate); nothing new is pub). No behavior change: workspace tests
(35), clippy -D warnings, and both Shiki compat lanes (91 + 88 tests)
pass unchanged.
