+++ crates/ferriki-core/src/lib.rs
@@ -1,18 +1,25 @@
 mod asset_catalog;
+mod grammar;
+mod injection;
+mod render;
+mod rule;
+mod scanner;
+mod theme;
+mod tokenize;
+mod types;
 
 use ferroni::regexec;
-use ferroni::scanner::{OnigString, Scanner, ScannerFindOptions};
 use napi::bindgen_prelude::*;
 use napi_derive::napi;
-use serde_json::{json, Value};
+use serde_json::Value;
 use std::cell::RefCell;
 use std::collections::{HashMap, HashSet};
-use std::hash::{BuildHasherDefault, Hasher};
-use std::sync::atomic::{AtomicU64, Ordering};
-use std::sync::Arc;
-use std::time::{Duration, Instant};
 
 pub use asset_catalog::{LanguageAssetCatalog, StandardAssetCatalogs, ThemeAssetCatalog};
+
+use grammar::*;
+use render::*;
+use types::*;
 
 #[napi(js_name = "ferrikiVersion")]
 pub fn ferriki_version() -> String {
@@ -29,5150 +36,6 @@ pub struct FerrikiHighlighter {
     compiled_grammars: RefCell<HashMap<String, CompiledGrammar>>,
     /// Maps target scope → list of injecting grammar scope names
     injection_map: RefCell<HashMap<String, Vec<String>>>,
-}
-
-#[derive(Clone, Copy)]
-enum LangMode {
-    Plaintext,
-    Json,
-    Grammar,
-}
-
-struct JsonToken {
-    kind: &'static str,
-    start_utf16: usize,
-    end_utf16: usize,
-    content: String,
-}
-
-#[derive(Clone)]
-struct StyledJsonToken {
-    content: String,
-    content_utf16_len: usize,
-    offset_utf16: usize,
-    color: Arc<str>,
-    font_style: u8,
-    dark_color: Option<Arc<str>>,
-}
-
-struct JsonThemeProfile {
-    pre_class: String,
-    pre_style: Option<String>,
-    theme_name: String,
-    fg: Option<String>,
-    bg: Option<String>,
-}
-
-struct HtmlThemeProfile {
-    pre_class: String,
-    pre_style: Option<String>,
-    theme_name: String,
-    dark_theme_name: Option<String>,
-    fg: Option<String>,
-    bg: Option<String>,
-    dark_fg: Option<String>,
-    dark_bg: Option<String>,
-    disable_token_coloring: bool,
-}
-
-struct ThemeRule {
-    scopes: Vec<String>,
-    /// Pre-split selector parts for each scope (avoids repeated split_whitespace)
-    scope_parts: Vec<Vec<String>>,
-    foreground: Option<Arc<str>>,
-    font_style: u8,
-}
-
-impl ThemeRule {
-    fn new(scopes: Vec<String>, foreground: Option<String>, font_style: u8) -> Self {
-        let scope_parts = scopes
-            .iter()
-            .map(|s| s.split_whitespace().map(str::to_owned).collect())
-            .collect();
-        // Pre-normalize foreground color at registration time
-        let foreground = foreground
-            .map(|c| normalize_hex_color(&c))
-            .map(Arc::<str>::from);
-        Self {
-            scopes,
-            scope_parts,
-            foreground,
-            font_style,
-        }
-    }
-}
-
-struct ThemeData {
-    name: String,
-    fg: String,
-    fg_normalized: Arc<str>,
-    bg: String,
-    settings: Vec<ThemeRule>,
-}
-
-struct GrammarRegistration {
-    scope_name: String,
-    grammar: Value,
-    aliases: Vec<String>,
-    has_explicit_grammar: bool,
-    inject_to: Vec<String>,
-}
-
-// ─────────────────────────────────────────────────────────────────────────────
-// New RuleId-based architecture (port of vscode-textmate)
-// ─────────────────────────────────────────────────────────────────────────────
-
-type RuleId = i32;
-const END_RULE_ID: RuleId = -1;
-static NEXT_ONIG_STR_ID: AtomicU64 = AtomicU64::new(1);
-
-#[derive(Clone)]
-struct GrammarCapture {
-    index: usize,
-    name: Option<String>,
-}
-
-struct RuleRegistry {
-    rules: Vec<Option<Rule>>,
-    next_id: RuleId,
-}
-
-impl RuleRegistry {
-    fn new() -> Self {
-        RuleRegistry {
-            rules: vec![None], // index 0 is unused (root sentinel)
-            next_id: 1,
-        }
-    }
-
-    fn alloc_id(&mut self) -> RuleId {
-        let id = self.next_id;
-        self.next_id += 1;
-        // Ensure vec is large enough
-        while self.rules.len() <= id as usize {
-            self.rules.push(None);
-        }
-        id
-    }
-
-    fn store(&mut self, id: RuleId, rule: Rule) {
-        self.rules[id as usize] = Some(rule);
-    }
-
-    fn get(&self, id: RuleId) -> Option<&Rule> {
-        if id < 0 || id as usize >= self.rules.len() {
-            return None;
-        }
-        self.rules[id as usize].as_ref()
-    }
-}
-
-enum Rule {
-    Match {
-        _id: RuleId,
-        name: Option<String>,
-        match_re: String,
-        captures: Vec<GrammarCapture>,
-    },
-    IncludeOnly {
-        _id: RuleId,
-        _name: Option<String>,
-        _content_name: Option<String>,
-        patterns: Vec<RuleId>,
-    },
-    BeginEnd {
-        _id: RuleId,
-        name: Option<String>,
-        content_name: Option<String>,
-        begin_re: String,
-        end_re: String,
-        end_has_back_references: bool,
-        apply_end_pattern_last: bool,
-        begin_captures: Vec<GrammarCapture>,
-        end_captures: Vec<GrammarCapture>,
-        patterns: Vec<RuleId>,
-    },
-    BeginWhile {
-        _id: RuleId,
-        name: Option<String>,
-        content_name: Option<String>,
-        begin_re: String,
-        while_re: String,
-        while_has_back_references: bool,
-        begin_captures: Vec<GrammarCapture>,
-        while_captures: Vec<GrammarCapture>,
-        patterns: Vec<RuleId>,
-    },
-}
-
-struct CompiledScanner {
-    scanner: Scanner,
-    rule_ids: Vec<RuleId>, // match index → RuleId
-    regexes: Vec<String>,
-    single_scanners: Vec<Option<Scanner>>,
-}
-
-struct StateFrame {
-    rule_id: RuleId,
-    _enter_pos: i32,
-    _anchor_pos: i32,
-    end_rule: Option<String>,
-    name_scopes: Vec<String>,
-    content_scopes: Vec<String>,
-}
-
-#[derive(Clone, Copy, PartialEq, Eq)]
-enum InjectionPriority {
-    Left,
-    Default,
-    Right,
-}
-
-struct CompiledSelector {
-    clauses: Vec<CompiledSelectorClause>,
-}
-
-struct CompiledSelectorClause {
-    disjuncts: Vec<CompiledSelectorDisjunct>,
-}
-
-struct CompiledSelectorDisjunct {
-    terms: Vec<CompiledSelectorTerm>,
-}
-
-struct CompiledSelectorTerm {
-    negate: bool,
-    expr: CompiledSelectorExpr,
-}
-
-enum CompiledSelectorExpr {
-    Token(String),
-    AnyOf(Vec<CompiledSelectorDisjunct>),
-}
-
-struct Injection {
-    compiled_selector: CompiledSelector,
-    rule_id: RuleId,
-    priority: InjectionPriority,
-}
-
-struct CompiledGrammar {
-    registry: RuleRegistry,
-    root_rule_id: RuleId,
-    injections: Vec<Injection>,
-    scanner_cache: HashMap<(RuleId, Option<String>), CompiledScanner>,
-    injection_scanner_cache: HashMap<RuleId, CompiledScanner>,
-    /// Cache for single-pattern while-condition scanners, keyed by regex string
-    while_scanner_cache: HashMap<String, Scanner>,
-}
-
-const COLOR_DEFAULT_FG: &str = "#DBD7CAEE";
-const COLOR_DEFAULT_BG: &str = "#121212";
-const COLOR_INHERIT: &str = "inherit";
-
-// ─────────────────────────────────────────────────────────────────────────────
-// Grammar parsing helpers (unchanged)
-// ─────────────────────────────────────────────────────────────────────────────
-
-fn parse_lang(options_json: &str) -> Option<String> {
-    let parsed: Value = serde_json::from_str(options_json).ok()?;
-    parsed.get("lang")?.as_str().map(str::to_owned)
-}
-
-fn parse_theme(options_json: &str) -> Option<String> {
-    let parsed: Value = serde_json::from_str(options_json).ok()?;
-    parsed.get("theme")?.as_str().map(str::to_owned)
-}
-
-fn parse_dual_themes(options_json: &str) -> Option<(String, String)> {
-    let parsed: Value = serde_json::from_str(options_json).ok()?;
-    let themes = parsed.get("themes")?.as_object()?;
-    let light = themes.get("light")?.as_str()?.to_owned();
-    let dark = themes.get("dark")?.as_str()?.to_owned();
-    Some((light, dark))
-}
-
-fn serialize_state_frames(stack: &[StateFrame], root_scope: Option<&str>) -> Value {
-    let mut obj = serde_json::Map::new();
-    if let Some(scope) = root_scope {
-        obj.insert("rootScope".to_owned(), Value::String(scope.to_owned()));
-    }
-    let frames: Vec<Value> = stack
-        .iter()
-        .map(|frame| {
-            json!({
-              "ruleId": frame.rule_id,
-              "endRule": frame.end_rule,
-              "nameScopes": frame.name_scopes,
-              "contentScopes": frame.content_scopes,
-            })
-        })
-        .collect();
-    obj.insert("frames".to_owned(), Value::Array(frames));
-    Value::Object(obj)
-}
-
-fn deserialize_state_frames(value: &Value) -> Option<Vec<StateFrame>> {
-    // Support both new format { rootScope, frames: [...] } and legacy bare array
-    let arr = if let Some(obj) = value.as_object() {
-        obj.get("frames")?.as_array()?
-    } else {
-        value.as_array()?
-    };
-    let mut frames = Vec::with_capacity(arr.len());
-    for item in arr {
-        let obj = item.as_object()?;
-        let rule_id = obj.get("ruleId")?.as_i64()? as RuleId;
-        let end_rule = obj
-            .get("endRule")
-            .and_then(|v| v.as_str().map(str::to_owned));
-        let name_scopes = obj
-            .get("nameScopes")
-            .and_then(|v| v.as_array())
-            .map(|arr| {
-                arr.iter()
-                    .filter_map(|v| v.as_str().map(str::to_owned))
-                    .collect()
-            })
-            .unwrap_or_default();
-        let content_scopes = obj
-            .get("contentScopes")
-            .and_then(|v| v.as_array())
-            .map(|arr| {
-                arr.iter()
-                    .filter_map(|v| v.as_str().map(str::to_owned))
-                    .collect()
-            })
-            .unwrap_or_default();
-        frames.push(StateFrame {
-            rule_id,
-            _enter_pos: 0,
-            _anchor_pos: 0,
-            end_rule,
-            name_scopes,
-            content_scopes,
-        });
-    }
-    Some(frames)
-}
-
-fn parse_initial_state_from_options(options_json: &str) -> Option<Vec<StateFrame>> {
-    let parsed: Value = serde_json::from_str(options_json).ok()?;
-    if let Some(rust_state) = parsed.get("_rustState") {
-        return deserialize_state_frames(rust_state);
-    }
-    None
-}
-
-fn parse_grammar_context_code(options_json: &str) -> Option<String> {
-    let parsed: Value = serde_json::from_str(options_json).ok()?;
-    parsed
-        .get("grammarContextCode")?
-        .as_str()
-        .map(str::to_owned)
-}
-
-fn parse_standard_asset_root(options_json: &str) -> Option<String> {
-    let parsed: Value = serde_json::from_str(options_json).ok()?;
-    parsed.get("standardAssetRoot")?.as_str().map(str::to_owned)
-}
-
-fn parse_string_array(value: Option<&Value>) -> Vec<String> {
-    let Some(Value::Array(items)) = value else {
-        return Vec::new();
-    };
-
-    items
-        .iter()
-        .filter_map(|item| item.as_str().map(str::to_owned))
-        .collect::<Vec<_>>()
-}
-
-fn parse_grammar_registration(payload_json: &str) -> Result<GrammarRegistration> {
-    let payload: Value = serde_json::from_str(payload_json).map_err(|err| {
-        Error::from_reason(format!("Failed to parse grammar registration JSON: {err}"))
-    })?;
-
-    let payload_obj = payload
-        .as_object()
-        .ok_or_else(|| Error::from_reason("Grammar registration payload must be an object."))?;
-
-    let has_explicit_grammar = payload_obj.contains_key("grammar")
-        || payload_obj.contains_key("patterns")
-        || payload_obj.contains_key("repository")
-        || payload_obj.contains_key("injections");
-    let mut grammar = payload
-        .get("grammar")
-        .cloned()
-        .unwrap_or_else(|| payload.clone());
-
-    let mut scope_name = payload_obj
-        .get("scopeName")
-        .and_then(Value::as_str)
-        .map(str::to_owned);
-    if scope_name.is_none() {
-        scope_name = grammar
-            .get("scopeName")
-            .and_then(Value::as_str)
-            .map(str::to_owned);
-    }
-    let scope_name = scope_name
-        .ok_or_else(|| Error::from_reason("Grammar registration requires `scopeName`."))?;
-
-    if let Value::Object(ref mut grammar_obj) = grammar {
-        grammar_obj
-            .entry("scopeName".to_owned())
-            .or_insert_with(|| Value::String(scope_name.clone()));
-    } else {
-        return Err(Error::from_reason(
-            "Grammar registration `grammar` must be an object.",
-        ));
-    }
-
-    let mut aliases = parse_string_array(payload_obj.get("aliases"));
-    if aliases.is_empty() {
-        aliases = parse_string_array(grammar.get("aliases"));
-    }
-
-    let inject_to = parse_string_array(payload_obj.get("injectTo"));
-
-    Ok(GrammarRegistration {
-        scope_name,
-        grammar,
-        aliases,
-        has_explicit_grammar,
-        inject_to,
-    })
-}
-
-fn parse_theme_registration(payload_json: &str) -> Result<ThemeData> {
-    let payload: Value = serde_json::from_str(payload_json).map_err(|err| {
-        Error::from_reason(format!("Failed to parse theme registration JSON: {err}"))
-    })?;
-
-    let obj = payload
-        .as_object()
-        .ok_or_else(|| Error::from_reason("Theme registration payload must be an object."))?;
-
-    let name = obj
-        .get("name")
-        .and_then(Value::as_str)
-        .ok_or_else(|| Error::from_reason("Theme registration requires `name`."))?
-        .to_owned();
-
-    let fg = obj
-        .get("fg")
-        .and_then(Value::as_str)
-        .unwrap_or("")
-        .to_owned();
-
-    let bg = obj
-        .get("bg")
-        .and_then(Value::as_str)
-        .unwrap_or("")
-        .to_owned();
-
-    let mut settings = Vec::new();
-    if let Some(Value::Array(rules)) = obj.get("settings") {
-        for rule in rules {
-            let rule_obj = match rule.as_object() {
-                Some(o) => o,
-                None => continue,
-            };
-            let scopes = match rule_obj.get("scope") {
-                Some(Value::Array(arr)) => arr
-                    .iter()
-                    .filter_map(|v| v.as_str().map(str::to_owned))
-                    .collect(),
-                _ => Vec::new(),
-            };
-            let foreground = rule_obj
-                .get("foreground")
-                .and_then(Value::as_str)
-                .map(str::to_owned);
-            let font_style = rule_obj
-                .get("fontStyle")
-                .and_then(Value::as_u64)
-                .unwrap_or(0) as u8;
-            settings.push(ThemeRule::new(scopes, foreground, font_style));
-        }
-    }
-
-    Ok(ThemeData {
-        name,
-        fg_normalized: Arc::<str>::from(normalize_hex_color(&fg)),
-        fg,
-        bg,
-        settings,
-    })
-}
-
-// ─────────────────────────────────────────────────────────────────────────────
-// Theme / scope resolution (unchanged)
-// ─────────────────────────────────────────────────────────────────────────────
-
-#[derive(Clone)]
-struct ResolvedStyle {
-    foreground: Option<Arc<str>>,
-    font_style: u8,
-}
-
-/// Identity hasher for already-hashed u64 cache keys.
-#[derive(Default)]
-struct U64IdentityHasher(u64);
-
-impl Hasher for U64IdentityHasher {
-    fn finish(&self) -> u64 {
-        self.0
-    }
-
-    fn write(&mut self, bytes: &[u8]) {
-        // Fallback path; u64 keys use `write_u64`.
-        let mut h: u64 = 0xcbf29ce484222325;
-        for b in bytes {
-            h ^= *b as u64;
-            h = h.wrapping_mul(0x100000001b3);
-        }
-        self.0 = h;
-    }
-
-    fn write_u64(&mut self, i: u64) {
-        self.0 = i;
-    }
-}
-
-type U64HashBuilder = BuildHasherDefault<U64IdentityHasher>;
-
-/// Cache for theme resolution results, keyed by a hash of the scope stack.
-struct ThemeCache {
-    map: HashMap<u64, ResolvedStyle, U64HashBuilder>,
-}
-
-impl ThemeCache {
-    fn new() -> Self {
-        Self {
-            map: HashMap::with_hasher(U64HashBuilder::default()),
-        }
-    }
-
-    /// Hash scope stack from &[String] without intermediate Vec<&str>
-    fn scope_hash_owned(scope_stack: &[String]) -> u64 {
-        let mut h: u64 = 0xcbf29ce484222325;
-        h ^= scope_stack.len() as u64;
-        h = h.wrapping_mul(0x100000001b3);
-        for s in scope_stack {
-            for b in s.as_bytes() {
-                h ^= *b as u64;
-                h = h.wrapping_mul(0x100000001b3);
-            }
-            h ^= 0xff;
-            h = h.wrapping_mul(0x100000001b3);
-        }
-        h
-    }
-
-    fn scope_hash_with_extra_owned(scope_stack: &[String], extra_scopes: &[String]) -> u64 {
-        let mut h: u64 = 0xcbf29ce484222325;
-        h ^= (scope_stack.len() + extra_scopes.len()) as u64;
-        h = h.wrapping_mul(0x100000001b3);
-        for s in scope_stack {
-            for b in s.as_bytes() {
-                h ^= *b as u64;
-                h = h.wrapping_mul(0x100000001b3);
-            }
-            h ^= 0xff;
-            h = h.wrapping_mul(0x100000001b3);
-        }
-        for scope in extra_scopes {
-            for b in scope.as_bytes() {
-                h ^= *b as u64;
-                h = h.wrapping_mul(0x100000001b3);
-            }
-            h ^= 0xfe;
-            h = h.wrapping_mul(0x100000001b3);
-        }
-        h
-    }
-
-    fn resolve_owned(&mut self, scope_stack: &[String], theme: &ThemeData) -> &ResolvedStyle {
-        let key = Self::scope_hash_owned(scope_stack);
-        self.map.entry(key).or_insert_with(|| {
-            let refs: Vec<&str> = scope_stack.iter().map(String::as_str).collect();
-            resolve_token_style(&refs, theme)
-        })
-    }
-
-    fn resolve_with_extra_owned(
-        &mut self,
-        scope_stack: &[String],
-        extra: &str,
-        theme: &ThemeData,
-    ) -> &ResolvedStyle {
-        let extra_scopes = parse_scope_list(extra);
-        let key = Self::scope_hash_with_extra_owned(scope_stack, &extra_scopes);
-        self.map.entry(key).or_insert_with(|| {
-            let mut refs: Vec<&str> = scope_stack.iter().map(String::as_str).collect();
-            refs.extend(extra_scopes.iter().map(String::as_str));
-            resolve_token_style(&refs, theme)
-        })
-    }
-}
-
-fn parse_scope_list(value: &str) -> Vec<String> {
-    value.split_whitespace().map(str::to_owned).collect()
-}
-
-fn scope_component_matches(selector: &str, scope: &str) -> bool {
-    selector == scope
-        || (scope.starts_with(selector) && scope.as_bytes().get(selector.len()) == Some(&b'.'))
-}
-
-fn selector_matches_presplit(parts: &[String], scope_stack: &[&str]) -> Option<usize> {
-    if parts.is_empty() {
-        return None;
-    }
-
-    let innermost = scope_stack.last()?;
-    if !scope_component_matches(&parts[parts.len() - 1], innermost) {
-        return None;
-    }
-
-    if parts.len() == 1 {
-        return Some(parts[0].len());
-    }
-
-    let mut part_idx = (parts.len() - 2) as isize;
-    let parent_scopes = &scope_stack[..scope_stack.len() - 1];
-    let mut stack_idx = (parent_scopes.len() as isize) - 1;
-
-    while part_idx >= 0 && stack_idx >= 0 {
-        if scope_component_matches(&parts[part_idx as usize], parent_scopes[stack_idx as usize]) {
-            part_idx -= 1;
-        }
-        stack_idx -= 1;
-    }
-
-    if part_idx < 0 {
-        return Some(parts.iter().map(|p| p.len()).sum());
-    }
-
-    None
-}
-
-fn resolve_token_style(scope_stack: &[&str], theme: &ThemeData) -> ResolvedStyle {
-    let mut best_score: usize = 0;
-    let mut best_fg: Option<Arc<str>> = None;
-    let mut best_font_style: u8 = 0;
-    let mut has_global = false;
-
-    for rule in &theme.settings {
-        if rule.scopes.is_empty() {
-            if !has_global {
-                has_global = true;
-                best_fg = rule.foreground.clone();
-                best_font_style = rule.font_style;
-            }
-            continue;
-        }
-
-        for parts in &rule.scope_parts {
-            if let Some(score) = selector_matches_presplit(parts, scope_stack) {
-                if score >= best_score {
-                    best_score = score;
-                    best_fg = rule.foreground.clone().or(best_fg);
-                    best_font_style = rule.font_style;
-                }
-            }
-        }
-    }
-
-    ResolvedStyle {
-        foreground: best_fg,
-        font_style: best_font_style,
-    }
-}
-
-// ─────────────────────────────────────────────────────────────────────────────
-// Grammar capture parsing (unchanged)
-// ─────────────────────────────────────────────────────────────────────────────
-
-fn parse_grammar_captures(value: Option<&Value>) -> Vec<GrammarCapture> {
-    let Some(Value::Object(obj)) = value else {
-        return Vec::new();
-    };
-
-    let mut captures = obj
-        .iter()
-        .filter_map(|(key, capture)| {
-            let index = key.parse::<usize>().ok()?;
-            let name = capture
-                .as_object()
-                .and_then(|entry| entry.get("name"))
-                .and_then(Value::as_str)
-                .map(str::to_owned);
-            Some(GrammarCapture { index, name })
-        })
-        .collect::<Vec<_>>();
-
-    captures.sort_by_key(|entry| entry.index);
-    captures
-}
-
-// ─────────────────────────────────────────────────────────────────────────────
-// Injection selector parsing (unchanged from old code)
-// ─────────────────────────────────────────────────────────────────────────────
-
-fn parse_injection_clause_priority(clause: &str) -> (InjectionPriority, &str) {
-    let mut priority = InjectionPriority::Default;
-    let mut rest = clause.trim();
-
-    loop {
-        if let Some(stripped) = rest.strip_prefix("L:") {
-            priority = InjectionPriority::Left;
-            rest = stripped.trim_start();
-            continue;
-        }
-        if let Some(stripped) = rest.strip_prefix("R:") {
-            priority = InjectionPriority::Right;
-            rest = stripped.trim_start();
-            continue;
-        }
-        if let Some(stripped) = rest.strip_prefix("B:") {
-            priority = InjectionPriority::Default;
-            rest = stripped.trim_start();
-            continue;
-        }
-        break;
-    }
-
-    (priority, rest.trim())
-}
-
-fn scope_token_matches_scope(token: &str, scope: &str) -> bool {
-    if token.is_empty() {
-        return false;
-    }
-    if token == "*" || token == "$self" || token == scope {
-        return true;
-    }
-    if scope
-        .strip_prefix(token)
-        .map(|rest| rest.is_empty() || rest.starts_with('.'))
-        .unwrap_or(false)
-    {
-        return true;
-    }
-    false
-}
-
-fn scope_token_matches_stack(token: &str, scope_stack: &[String]) -> bool {
-    scope_stack
-        .iter()
-        .any(|scope| scope_token_matches_scope(token, scope))
-}
-
-fn parse_injection_term(term: &str) -> (bool, &str) {
-    let mut negate = false;
-    let mut out = term.trim();
-
-    loop {
-        if let Some(rest) = out.strip_prefix('!') {
-            negate = !negate;
-            out = rest.trim_start();
-            continue;
-        }
-        if let Some(rest) = out.strip_prefix('-') {
-            negate = !negate;
-            out = rest.trim_start();
-            continue;
-        }
-        break;
-    }
-
-    let out = out.trim_matches(|ch: char| matches!(ch, '(' | ')' | '^'));
-    (negate, out)
-}
-
-fn split_top_level<'a>(input: &'a str, separators: &[char]) -> Vec<&'a str> {
-    let mut out = Vec::new();
-    let mut depth = 0usize;
-    let mut segment_start = 0usize;
-
-    for (idx, ch) in input.char_indices() {
-        match ch {
-            '(' => depth = depth.saturating_add(1),
-            ')' => depth = depth.saturating_sub(1),
-            _ => {}
-        }
-
-        if depth == 0 && separators.contains(&ch) {
-            if segment_start <= idx {
-                let segment = input[segment_start..idx].trim();
-                if !segment.is_empty() {
-                    out.push(segment);
-                }
-            }
-            segment_start = idx.saturating_add(ch.len_utf8());
-        }
-    }
-
-    if segment_start <= input.len() {
-        let segment = input[segment_start..].trim();
-        if !segment.is_empty() {
-            out.push(segment);
-        }
-    }
-
-    out
-}
-
-fn split_selector_terms(disjunct: &str) -> Vec<&str> {
-    let mut out = Vec::new();
-    let mut depth = 0usize;
-    let mut segment_start: Option<usize> = None;
-
-    for (idx, ch) in disjunct.char_indices() {
-        match ch {
-            '(' => {
-                depth = depth.saturating_add(1);
-                if segment_start.is_none() {
-                    segment_start = Some(idx);
-                }
-            }
-            ')' => depth = depth.saturating_sub(1),
-            '&' if depth == 0 => {
-                if let Some(start) = segment_start.take() {
-                    let term = disjunct[start..idx].trim();
-                    if !term.is_empty() {
-                        out.push(term);
-                    }
-                }
-            }
-            _ if depth == 0 && ch.is_whitespace() => {
-                if let Some(start) = segment_start.take() {
-                    let term = disjunct[start..idx].trim();
-                    if !term.is_empty() {
-                        out.push(term);
-                    }
-                }
-            }
-            _ => {
-                if segment_start.is_none() {
-                    segment_start = Some(idx);
-                }
-            }
-        }
-    }
-
-    if let Some(start) = segment_start {
-        let term = disjunct[start..].trim();
-        if !term.is_empty() {
-            out.push(term);
-        }
-    }
-
-    out
-}
-
-fn compile_selector_disjunct(disjunct: &str) -> Option<CompiledSelectorDisjunct> {
-    let raw_terms = split_selector_terms(disjunct);
-    if raw_terms.is_empty() {
-        return None;
-    }
-
-    let mut terms: Vec<CompiledSelectorTerm> = Vec::new();
-    let mut index = 0usize;
-    while index < raw_terms.len() {
-        let mut raw_term = raw_terms[index].trim();
-        if raw_term.is_empty() {
-            index = index.saturating_add(1);
-            continue;
-        }
-
-        let mut detached_negate = false;
-        if (raw_term == "!" || raw_term == "-") && index + 1 < raw_terms.len() {
-            detached_negate = true;
-            index = index.saturating_add(1);
-            raw_term = raw_terms[index].trim();
-        }
-
-        let (term_negate, term) = parse_injection_term(raw_term);
-        let negate = detached_negate ^ term_negate;
-        if term.is_empty() {
-            index = index.saturating_add(1);
-            continue;
-        }
-
-        let branches = split_top_level(term, &['|']);
-        let expr = if branches.len() > 1 {
-            let mut compiled_branches: Vec<CompiledSelectorDisjunct> = Vec::new();
-            for branch in branches {
-                if let Some(compiled) = compile_selector_disjunct(branch) {
-                    compiled_branches.push(compiled);
-                }
-            }
-            if compiled_branches.is_empty() {
-                index = index.saturating_add(1);
-                continue;
-            }
-            CompiledSelectorExpr::AnyOf(compiled_branches)
-        } else {
-            CompiledSelectorExpr::Token(term.to_owned())
-        };
-
-        terms.push(CompiledSelectorTerm { negate, expr });
-
-        index = index.saturating_add(1);
-    }
-
-    if terms.is_empty() {
-        None
-    } else {
-        Some(CompiledSelectorDisjunct { terms })
-    }
-}
-
-fn compile_selector(selector: &str) -> CompiledSelector {
-    let mut clauses: Vec<CompiledSelectorClause> = Vec::new();
-
-    for clause in split_top_level(selector, &[',']) {
-        let (_priority, normalized) = parse_injection_clause_priority(clause);
-        if normalized.is_empty() {
-            continue;
-        }
-
-        let mut disjuncts: Vec<CompiledSelectorDisjunct> = Vec::new();
-        for disjunct in split_top_level(normalized, &['|']) {
-            if let Some(compiled) = compile_selector_disjunct(disjunct) {
-                disjuncts.push(compiled);
-            }
-        }
-        if !disjuncts.is_empty() {
-            clauses.push(CompiledSelectorClause { disjuncts });
-        }
-    }
-
-    CompiledSelector { clauses }
-}
-
-fn selector_disjunct_matches_compiled(
-    disjunct: &CompiledSelectorDisjunct,
-    scope_stack: &[String],
-) -> bool {
-    let mut has_term = false;
-    for term in &disjunct.terms {
-        has_term = true;
-        let matches = match &term.expr {
-            CompiledSelectorExpr::Token(token) => scope_token_matches_stack(token, scope_stack),
-            CompiledSelectorExpr::AnyOf(branches) => branches
-                .iter()
-                .any(|branch| selector_disjunct_matches_compiled(branch, scope_stack)),
-        };
-
-        if term.negate {
-            if matches {
-                return false;
-            }
-        } else if !matches {
-            return false;
-        }
-    }
-
-    has_term
-}
-
-fn selector_matches_compiled(selector: &CompiledSelector, scope_stack: &[String]) -> bool {
-    selector.clauses.iter().any(|clause| {
-        clause
-            .disjuncts
-            .iter()
-            .any(|disjunct| selector_disjunct_matches_compiled(disjunct, scope_stack))
-    })
-}
-
-// ─────────────────────────────────────────────────────────────────────────────
-// Rule compilation (port of vscode-textmate getCompiledRuleId/_compilePatterns)
-// ─────────────────────────────────────────────────────────────────────────────
-
-/// Port of vscode-textmate initGrammar(): creates synthetic $self and $base
-/// entries in the repository.
-/// If `base_grammar` is provided, $base points to the base grammar's $self.
-/// Otherwise, $base = $self (the grammar itself).
-fn init_grammar(grammar: &Value, base_grammar: Option<&Value>) -> Value {
-    let mut g = grammar.clone();
-    if let Value::Object(ref mut obj) = g {
-        // Ensure repository exists
-        if !obj.contains_key("repository") {
-            obj.insert("repository".to_owned(), json!({}));
-        }
-
-        // Build $self = { patterns: grammar.patterns, name: grammar.scopeName }
-        let self_entry = {
-            let mut entry = serde_json::Map::new();
-            if let Some(patterns) = obj.get("patterns").cloned() {
-                entry.insert("patterns".to_owned(), patterns);
-            }
-            if let Some(name) = obj.get("scopeName").cloned() {
-                entry.insert("name".to_owned(), name);
-            }
-            Value::Object(entry)
-        };
-
-        // Build $base entry: from base_grammar if provided, else same as $self
-        let base_entry = if let Some(base) = base_grammar {
-            if let Some(base_obj) = base.as_object() {
-                let mut entry = serde_json::Map::new();
-                if let Some(patterns) = base_obj.get("patterns").cloned() {
-                    entry.insert("patterns".to_owned(), patterns);
-                }
-                if let Some(name) = base_obj.get("scopeName").cloned() {
-                    entry.insert("name".to_owned(), name);
-                }
-                // Also merge in the base grammar's repository for $base includes
-                if let Some(repo) = base_obj.get("repository").cloned() {
-                    entry.insert("repository".to_owned(), repo);
-                }
-                Value::Object(entry)
-            } else {
-                self_entry.clone()
-            }
-        } else {
-            self_entry.clone()
-        };
-
-        if let Some(Value::Object(ref mut repo)) = obj.get_mut("repository") {
-            repo.insert("$self".to_owned(), self_entry);
-            repo.insert("$base".to_owned(), base_entry);
-        }
-    }
-    g
-}
-
-/// Check if a pattern string contains back-references like \1, \2, etc.
-fn has_back_references(pattern: &str) -> bool {
-    let mut chars = pattern.chars();
-    while let Some(ch) = chars.next() {
-        if ch == '\\' {
-            match chars.next() {
-                Some(next) if next.is_ascii_digit() && next != '0' => return true,
-                Some(_) => {}
-                None => break,
-            }
-        }
-    }
-    false
-}
-
-/// Compile a grammar descriptor into the rule registry, returning its RuleId.
-/// Port of vscode-textmate RuleFactory.getCompiledRuleId().
-fn compile_rule(
-    desc: &Value,
-    registry: &mut RuleRegistry,
-    compiled_map: &mut HashMap<String, RuleId>,
-    repository: &serde_json::Map<String, Value>,
-    grammar_pool: &HashMap<String, Value>,
-    desc_key: &str,
-    host_grammar: Option<&Value>,
-) -> Option<RuleId> {
-    compile_rule_inner(
-        desc,
-        registry,
-        compiled_map,
-        repository,
-        grammar_pool,
-        desc_key,
-        host_grammar,
-        0,
-    )
-}
-
-const MAX_COMPILE_DEPTH: usize = 64;
-
-#[allow(clippy::too_many_arguments)]
-fn compile_rule_inner(
-    desc: &Value,
-    registry: &mut RuleRegistry,
-    compiled_map: &mut HashMap<String, RuleId>,
-    repository: &serde_json::Map<String, Value>,
-    grammar_pool: &HashMap<String, Value>,
-    desc_key: &str,
-    host_grammar: Option<&Value>,
-    depth: usize,
-) -> Option<RuleId> {
-    if depth > MAX_COMPILE_DEPTH {
-        return None;
-    }
-
-    // Memoize: if already compiled, return existing id
-    if let Some(&id) = compiled_map.get(desc_key) {
-        return Some(id);
-    }
-
-    let obj = desc.as_object()?;
-
-    // Allocate ID before recursing (prevents infinite recursion)
-    let id = registry.alloc_id();
-    compiled_map.insert(desc_key.to_owned(), id);
-
-    let name = obj.get("name").and_then(Value::as_str).map(str::to_owned);
-    let content_name = obj
-        .get("contentName")
-        .and_then(Value::as_str)
-        .map(str::to_owned);
-
-    if let Some(match_re) = obj.get("match").and_then(Value::as_str) {
-        // MatchRule
-        let captures = parse_grammar_captures(obj.get("captures"));
-        registry.store(
-            id,
-            Rule::Match {
-                _id: id,
-                name,
-                match_re: match_re.to_owned(),
-                captures,
-            },
-        );
-        return Some(id);
-    }
-
-    if let Some(begin_re) = obj.get("begin").and_then(Value::as_str) {
-        let while_re = obj.get("while").and_then(Value::as_str);
-
-        if let Some(while_re_str) = while_re {
-            // BeginWhileRule
-            let captures = parse_grammar_captures(obj.get("captures"));
-            let mut begin_captures = parse_grammar_captures(obj.get("beginCaptures"));
-            if begin_captures.is_empty() {
-                begin_captures = captures.clone();
-            }
-            let mut while_captures = parse_grammar_captures(obj.get("whileCaptures"));
-            if while_captures.is_empty() {
-                while_captures = captures;
-            }
-            let patterns = compile_patterns_inner(
-                obj.get("set").or_else(|| obj.get("patterns")),
-                registry,
-                compiled_map,
-                repository,
-                grammar_pool,
-                desc_key,
-                host_grammar,
-                depth,
-            );
-            registry.store(
-                id,
-                Rule::BeginWhile {
-                    _id: id,
-                    name,
-                    content_name,
-                    begin_re: begin_re.to_owned(),
-                    while_re: while_re_str.to_owned(),
-                    while_has_back_references: has_back_references(while_re_str),
-                    begin_captures,
-                    while_captures,
-                    patterns,
-                },
-            );
-            return Some(id);
-        }
-
-        // BeginEndRule
-        let end_re_str = obj.get("end").and_then(Value::as_str).unwrap_or("\u{FFFF}");
-        let captures = parse_grammar_captures(obj.get("captures"));
-        let mut begin_captures = parse_grammar_captures(obj.get("beginCaptures"));
-        if begin_captures.is_empty() {
-            begin_captures = captures.clone();
-        }
-        let mut end_captures = parse_grammar_captures(obj.get("endCaptures"));
-        if end_captures.is_empty() {
-            end_captures = captures;
-        }
-        let apply_end_pattern_last = obj
-            .get("applyEndPatternLast")
-            .and_then(|v| v.as_bool().or_else(|| v.as_u64().map(|n| n != 0)))
-            .unwrap_or(false);
-        let patterns = compile_patterns_inner(
-            obj.get("set").or_else(|| obj.get("patterns")),
-            registry,
-            compiled_map,
-            repository,
-            grammar_pool,
-            desc_key,
-            host_grammar,
-            depth,
-        );
-        registry.store(
-            id,
-            Rule::BeginEnd {
-                _id: id,
-                name,
-                content_name,
-                begin_re: begin_re.to_owned(),
-                end_re: end_re_str.to_owned(),
-                end_has_back_references: has_back_references(end_re_str),
-                apply_end_pattern_last,
-                begin_captures,
-                end_captures,
-                patterns,
-            },
-        );
-        return Some(id);
-    }
-
-    // IncludeOnlyRule: has patterns (or is a bare include wrapper)
-    // If it has an include directive, wrap it in patterns
-    let nested_patterns = if obj.contains_key("include") {
-        // Bare include — treat as a single-element patterns list
-        compile_patterns_inner(
-            Some(&json!([desc.clone()])),
-            registry,
-            compiled_map,
-            repository,
-            grammar_pool,
-            desc_key,
-            host_grammar,
-            depth,
-        )
-    } else {
-        compile_patterns_inner(
-            obj.get("patterns").or_else(|| obj.get("set")),
-            registry,
-            compiled_map,
-            repository,
-            grammar_pool,
-            desc_key,
-            host_grammar,
-            depth,
-        )
-    };
-
-    registry.store(
-        id,
-        Rule::IncludeOnly {
-            _id: id,
-            _name: name,
-            _content_name: content_name,
-            patterns: nested_patterns,
-        },
-    );
-    Some(id)
-}
-
-/// Stable identifier for a repository, using the pointer address of the Map.
-/// This ensures that the same repo key name in different grammars gets a distinct cache key.
-fn repo_id(repository: &serde_json::Map<String, Value>) -> usize {
-    repository as *const _ as usize
-}
-
-#[allow(clippy::too_many_arguments)]
-fn compile_patterns_inner(
-    patterns: Option<&Value>,
-    registry: &mut RuleRegistry,
-    compiled_map: &mut HashMap<String, RuleId>,
-    repository: &serde_json::Map<String, Value>,
-    grammar_pool: &HashMap<String, Value>,
-    parent_key: &str,
-    host_grammar: Option<&Value>,
-    depth: usize,
-) -> Vec<RuleId> {
-    let Some(Value::Array(items)) = patterns else {
-        return Vec::new();
-    };
-
-    let mut out = Vec::new();
-    for (idx, item) in items.iter().enumerate() {
-        let Some(obj) = item.as_object() else {
-            continue;
-        };
-
-        if let Some(include) = obj.get("include").and_then(Value::as_str) {
-            // Handle include references
-            if include == "$self" || include == "$base" {
-                let key = include;
-                if let Some(target) = repository.get(key) {
-                    // Use stable key scoped to this repository to enable memoization across call sites
-                    let desc_key = format!("{}:repo/{key}", repo_id(repository));
-                    if let Some(rule_id) = compile_rule_inner(
-                        target,
-                        registry,
-                        compiled_map,
-                        repository,
-                        grammar_pool,
-                        &desc_key,
-                        host_grammar,
-                        depth + 1,
-                    ) {
-                        out.push(rule_id);
-                    }
-                }
-                continue;
-            }
-
-            if let Some(key) = include.strip_prefix('#') {
-                if let Some(target) = repository.get(key) {
-                    // Use stable key scoped to this repository to enable memoization.
-                    // This is critical for grammars with cycles (e.g. jsx-children ↔ jsx-tag).
-                    let desc_key = format!("{}:repo/{key}", repo_id(repository));
-                    if let Some(rule_id) = compile_rule_inner(
-                        target,
-                        registry,
-                        compiled_map,
-                        repository,
-                        grammar_pool,
-                        &desc_key,
-                        host_grammar,
-                        depth + 1,
-                    ) {
-                        out.push(rule_id);
-                    }
-                }
-                continue;
-            }
-
-            // Cross-grammar reference: "scope#key"
-            if let Some((scope, key)) = include.split_once('#') {
-                if !scope.is_empty() && !key.is_empty() {
-                    if let Some(target_grammar) = grammar_pool.get(scope) {
-                        // Pass the host grammar as $base for the target grammar
-                        let initialized = init_grammar(target_grammar, host_grammar);
-                        if let Some(target_obj) = initialized.as_object() {
-                            if let Some(target_repo) =
-                                target_obj.get("repository").and_then(Value::as_object)
-                            {
-                                if let Some(target_rule) = target_repo.get(key) {
-                                    let desc_key = format!("cross:{scope}#{key}");
-                                    if let Some(rule_id) = compile_rule_inner(
-                                        target_rule,
-                                        registry,
-                                        compiled_map,
-                                        target_repo,
-                                        grammar_pool,
-                                        &desc_key,
-                                        host_grammar,
-                                        depth + 1,
-                                    ) {
-                                        out.push(rule_id);
-                                    }
-                                }
-                            }
-                        }
-                    }
-                }
-                continue;
-            }
-
-            // Top-level scope reference: "source.xxx"
-            if let Some(target_grammar) = grammar_pool.get(include) {
-                // Pass the host grammar as $base for the target grammar
-                let initialized = init_grammar(target_grammar, host_grammar);
-                if let Some(target_obj) = initialized.as_object() {
-                    if let Some(target_repo) =
-                        target_obj.get("repository").and_then(Value::as_object)
-                    {
-                        if let Some(self_entry) = target_repo.get("$self") {
-                            let desc_key = format!("scope:{include}/$self");
-                            if let Some(rule_id) = compile_rule_inner(
-                                self_entry,
-                                registry,
-                                compiled_map,
-                                target_repo,
-                                grammar_pool,
-                                &desc_key,
-                                host_grammar,
-                                depth + 1,
-                            ) {
-                                out.push(rule_id);
-                            }
-                        }
-                    }
-                }
-            }
-            continue;
-        }
-
-        // Normal rule (not an include)
-        let desc_key = format!("{parent_key}/pat/{idx}");
-        if let Some(rule_id) = compile_rule_inner(
-            item,
-            registry,
-            compiled_map,
-            repository,
-            grammar_pool,
-            &desc_key,
-            host_grammar,
-            depth + 1,
-        ) {
-            out.push(rule_id);
-        }
-    }
-
-    out
-}
-
-// ─────────────────────────────────────────────────────────────────────────────
-// Scanner compilation (port of Rule.collectPatterns / _getCachedCompiledPatterns)
-// ─────────────────────────────────────────────────────────────────────────────
-
-/// Maximum number of regex patterns allowed per Scanner.
-const MAX_SCANNER_PATTERNS: usize = 256;
-
-/// Recursively collect (regex, rule_id) pairs from a rule's patterns.
-fn collect_patterns(rule_id: RuleId, registry: &RuleRegistry, out: &mut Vec<(String, RuleId)>) {
-    let Some(rule) = registry.get(rule_id) else {
-        return;
-    };
-
-    match rule {
-        Rule::Match { match_re, _id, .. } => {
-            out.push((match_re.clone(), *_id));
-        }
-        Rule::IncludeOnly { patterns, .. } => {
-            for &pat_id in patterns {
-                if out.len() >= MAX_SCANNER_PATTERNS {
-                    break;
-                }
-                collect_patterns(pat_id, registry, out);
-            }
-        }
-        Rule::BeginEnd { begin_re, _id, .. } => {
-            out.push((begin_re.clone(), *_id));
-        }
-        Rule::BeginWhile { begin_re, _id, .. } => {
-            out.push((begin_re.clone(), *_id));
-        }
-    }
-}
-
-/// Build a compiled scanner for matching against a rule's child patterns,
-/// optionally including an end pattern.
-fn build_scanner_for_rule(
-    rule_id: RuleId,
-    registry: &RuleRegistry,
-    end_re: Option<&str>,
-    apply_end_pattern_last: bool,
-) -> Option<CompiledScanner> {
-    let mut pattern_pairs: Vec<(String, RuleId)> = Vec::new();
-
-    // Collect child patterns
-    let rule = registry.get(rule_id)?;
-
-    let child_patterns = match rule {
-        Rule::IncludeOnly { patterns, .. } => patterns.clone(),
-        Rule::BeginEnd { patterns, .. } => patterns.clone(),
-        Rule::BeginWhile { patterns, .. } => patterns.clone(),
-        Rule::Match { .. } => Vec::new(),
-    };
-
-    // Insert end pattern (if any) at beginning or end based on applyEndPatternLast
-    if !apply_end_pattern_last {
-        if let Some(end) = end_re {
-            pattern_pairs.push((end.to_owned(), END_RULE_ID));
-        }
-    }
-
-    for &pat_id in &child_patterns {
-        if pattern_pairs.len() >= MAX_SCANNER_PATTERNS {
-            break;
-        }
-        collect_patterns(pat_id, registry, &mut pattern_pairs);
-    }
-
-    if apply_end_pattern_last {
-        if let Some(end) = end_re {
-            if pattern_pairs.len() < MAX_SCANNER_PATTERNS {
-                pattern_pairs.push((end.to_owned(), END_RULE_ID));
-            }
-        }
-    }
-
-    if pattern_pairs.is_empty() {
-        return None;
-    }
-
-    // Build scanner, filtering out invalid regexes
-    let regexes: Vec<String> = pattern_pairs.iter().map(|(re, _)| re.clone()).collect();
-    let rule_ids: Vec<RuleId> = pattern_pairs.iter().map(|(_, id)| *id).collect();
-
-    if regexes.len() > 128 {
-        // For large pattern sets, try building the scanner directly
-        let regex_refs: Vec<&str> = regexes.iter().map(String::as_str).collect();
-        match Scanner::new(&regex_refs) {
-            Ok(scanner) => {
-                let single_scanners = std::iter::repeat_with(|| None)
-                    .take(regexes.len())
-                    .collect();
-                Some(CompiledScanner {
-                    scanner,
-                    rule_ids,
-                    regexes,
-                    single_scanners,
-                })
-            }
-            Err(_) => None,
-        }
-    } else {
-        // For smaller sets, validate each regex individually
-        let mut valid_regexes = Vec::new();
-        let mut valid_ids = Vec::new();
-        for (regex, rule_id) in regexes.into_iter().zip(rule_ids) {
-            if Scanner::new(&[regex.as_str()]).is_ok() {
-                valid_regexes.push(regex);
-                valid_ids.push(rule_id);
-            }
-        }
-
-        if valid_regexes.is_empty() {
-            return None;
-        }
-
-        let regex_refs: Vec<&str> = valid_regexes.iter().map(String::as_str).collect();
-        match Scanner::new(&regex_refs) {
-            Ok(scanner) => {
-                let single_scanners = std::iter::repeat_with(|| None)
-                    .take(valid_regexes.len())
-                    .collect();
-                Some(CompiledScanner {
-                    scanner,
-                    rule_ids: valid_ids,
-                    regexes: valid_regexes,
-                    single_scanners,
-                })
-            }
-            Err(_) => None,
-        }
-    }
-}
-
-fn find_next_match_ordered(
-    compiled_scanner: &mut CompiledScanner,
-    input: &OnigString,
-    line_str_id: u64,
-    cursor: usize,
-    find_options: ScannerFindOptions,
-) -> Option<ferroni::scanner::ScannerMatch> {
-    let best = compiled_scanner.scanner.find_next_match_utf16_with_id(
-        input,
-        line_str_id,
-        cursor,
-        find_options,
-    )?;
-    let mut best_match = best;
-    let mut best_start = best_match
-        .capture_indices
-        .first()
-        .map(|capture| capture.start)
-        .unwrap_or(usize::MAX);
-
-    for index in 0..best_match.index {
-        if compiled_scanner.single_scanners[index].is_none() {
-            compiled_scanner.single_scanners[index] =
-                Scanner::new(&[compiled_scanner.regexes[index].as_str()]).ok();
-        }
-        let Some(scanner) = compiled_scanner.single_scanners[index].as_mut() else {
-            continue;
-        };
-        let Some(mut candidate) =
-            scanner.find_next_match_utf16_with_id(input, line_str_id, cursor, find_options)
-        else {
-            continue;
-        };
-        candidate.index = index;
-        let candidate_start = candidate
-            .capture_indices
-            .first()
-            .map(|capture| capture.start)
-            .unwrap_or(usize::MAX);
-        // On a tie in start position the lower pattern index wins (TextMate
-        // declaration order); candidates are visited in ascending index order.
-        if candidate_start < best_start
-            || (candidate_start == best_start && candidate.index < best_match.index)
-        {
-            best_start = candidate_start;
-            best_match = candidate;
-            if best_start == cursor {
-                break;
-            }
-        }
-    }
-
-    Some(best_match)
-}
-
-// ─────────────────────────────────────────────────────────────────────────────
-// Injection compilation
-// ─────────────────────────────────────────────────────────────────────────────
-
-fn collect_injections(
-    grammar: &Value,
-    registry: &mut RuleRegistry,
-    compiled_map: &mut HashMap<String, RuleId>,
-    repository: &serde_json::Map<String, Value>,
-    grammar_pool: &HashMap<String, Value>,
-) -> Vec<Injection> {
-    let Some(obj) = grammar.as_object() else {
-        return Vec::new();
-    };
-
-    let Some(injections) = obj.get("injections").and_then(Value::as_object) else {
-        return Vec::new();
-    };
-
-    let mut result = Vec::new();
-
-    for (selector, injection_value) in injections {
-        let compiled_selector = compile_selector(selector);
-        if compiled_selector.clauses.is_empty() {
-            continue;
-        }
-
-        let (priority, _normalized) = parse_injection_clause_priority(selector);
-
-        // Compile the injection rule
-        let desc_key = format!("injection:{selector}");
-        if let Some(rule_id) = compile_rule(
-            injection_value,
-            registry,
-            compiled_map,
-            repository,
-            grammar_pool,
-            &desc_key,
-            None,
-        ) {
-            result.push(Injection {
-                compiled_selector,
-                rule_id,
-                priority,
-            });
-        }
-    }
-
-    result
-}
-
-/// Collect an external injection from a grammar that declares `injectTo` targeting
-/// the grammar being compiled. Reads the `injectionSelector` from the external
-/// grammar's top-level field and compiles the external grammar's root patterns
-/// as an injection rule.
-fn collect_external_injection(
-    ext_grammar: &Value,
-    injections: &mut Vec<Injection>,
-    registry: &mut RuleRegistry,
-    compiled_map: &mut HashMap<String, RuleId>,
-    _repository: &serde_json::Map<String, Value>,
-    grammar_pool: &HashMap<String, Value>,
-) {
-    let ext_obj = match ext_grammar.as_object() {
-        Some(obj) => obj,
-        None => return,
-    };
-
-    // The injectionSelector is a top-level field on the grammar JSON
-    let selector = match ext_obj.get("injectionSelector").and_then(Value::as_str) {
-        Some(s) if !s.is_empty() => s.to_owned(),
-        _ => return,
-    };
-    let compiled_selector = compile_selector(&selector);
-    if compiled_selector.clauses.is_empty() {
-        return;
-    }
-
-    let (priority, _normalized) = parse_injection_clause_priority(&selector);
-
-    let ext_scope = ext_obj
-        .get("scopeName")
-        .and_then(Value::as_str)
-        .unwrap_or("");
-    let desc_key = format!("external-injection:{ext_scope}");
-
-    // Build the external grammar's patterns as an injection rule.
-    // We initialize the external grammar (which merges repository/$base/$self)
-    // and compile its $self entry as the injection rule.
-    let initialized = init_grammar(ext_grammar, None);
-    let init_obj = match initialized.as_object() {
-        Some(obj) => obj,
-        None => return,
-    };
-    let ext_repo = match init_obj.get("repository").and_then(Value::as_object) {
-        Some(r) => r,
-        None => return,
-    };
-    let self_entry = match ext_repo.get("$self") {
-        Some(e) => e,
-        None => return,
-    };
-
-    if let Some(rule_id) = compile_rule(
-        self_entry,
-        registry,
-        compiled_map,
-        ext_repo,
-        grammar_pool,
-        &desc_key,
-        Some(ext_grammar),
-    ) {
-        injections.push(Injection {
-            compiled_selector,
-            rule_id,
-            priority,
-        });
-    }
-}
-
-// ─────────────────────────────────────────────────────────────────────────────
-// Grammar compilation (caching entry point)
-// ─────────────────────────────────────────────────────────────────────────────
-
-fn compile_grammar(
-    grammar: &Value,
-    grammar_pool: &HashMap<String, Value>,
-    injection_map: &HashMap<String, Vec<String>>,
-) -> Result<CompiledGrammar> {
-    let initialized = init_grammar(grammar, None);
-    let obj = initialized
-        .as_object()
-        .ok_or_else(|| Error::from_reason("Grammar is not an object"))?;
-    let repository = obj
-        .get("repository")
-        .and_then(Value::as_object)
-        .ok_or_else(|| Error::from_reason("Grammar has no repository"))?;
-
-    let mut registry = RuleRegistry::new();
-    let mut compiled_map: HashMap<String, RuleId> = HashMap::new();
-
-    let self_entry = repository
-        .get("$self")
-        .ok_or_else(|| Error::from_reason("Grammar missing $self after init"))?;
-    let root_rule_id = compile_rule(
-        self_entry,
-        &mut registry,
-        &mut compiled_map,
-        repository,
-        grammar_pool,
-        "root/$self",
-        Some(grammar),
-    )
-    .ok_or_else(|| Error::from_reason("Failed to compile root grammar rule"))?;
-
-    let mut injections = collect_injections(
-        &initialized,
-        &mut registry,
-        &mut compiled_map,
-        repository,
-        grammar_pool,
-    );
-
-    // Collect external injections from the injection map
-    let scope_name = grammar
-        .get("scopeName")
-        .and_then(Value::as_str)
-        .unwrap_or("");
-    if !scope_name.is_empty() {
-        let scope_parts: Vec<&str> = scope_name.split('.').collect();
-        for i in 1..=scope_parts.len() {
-            let sub_scope = scope_parts[..i].join(".");
-            if let Some(injecting_scopes) = injection_map.get(&sub_scope) {
-                for injecting_scope in injecting_scopes {
-                    if let Some(ext_grammar) = grammar_pool.get(injecting_scope) {
-                        collect_external_injection(
-                            ext_grammar,
-                            &mut injections,
-                            &mut registry,
-                            &mut compiled_map,
-                            repository,
-                            grammar_pool,
-                        );
-                    }
-                }
-            }
-        }
-    }
-
-    let mut scanner_cache: HashMap<(RuleId, Option<String>), CompiledScanner> = HashMap::new();
-    let root_scanner = build_scanner_for_rule(root_rule_id, &registry, None, false);
-    if let Some(scanner) = root_scanner {
-        scanner_cache.insert((root_rule_id, None), scanner);
-    }
-
-    Ok(CompiledGrammar {
-        registry,
-        root_rule_id,
-        injections,
-        scanner_cache,
-        injection_scanner_cache: HashMap::new(),
-        while_scanner_cache: HashMap::new(),
-    })
-}
-
-// ─────────────────────────────────────────────────────────────────────────────
-// Tokenization helpers (mostly unchanged)
-// ─────────────────────────────────────────────────────────────────────────────
-
-fn push_styled_slice(
-    out: &mut Vec<StyledJsonToken>,
-    utf16_map: &[usize],
-    code: &str,
-    start_utf16: usize,
-    end_utf16: usize,
-    color: &Arc<str>,
-    font_style: u8,
-) -> Result<()> {
-    if end_utf16 <= start_utf16 || end_utf16 >= utf16_map.len() {
-        return Ok(());
-    }
-
-    let start_byte = utf16_map[start_utf16];
-    let end_byte = utf16_map[end_utf16];
-    let content = code.get(start_byte..end_byte).ok_or_else(|| {
-        Error::from_reason("Ferriki grammar tokenizer failed to slice source text.")
-    })?;
-
-    if content.is_empty() {
-        return Ok(());
-    }
-
-    let utf16_len = end_utf16 - start_utf16;
-    out.push(StyledJsonToken {
-        content: content.to_owned(),
-        content_utf16_len: utf16_len,
-        offset_utf16: start_utf16,
-        color: color.clone(),
-        font_style,
-        dark_color: None,
-    });
-    Ok(())
-}
-
-struct CaptureRange {
-    start: usize,
-    end: usize,
-    color: Arc<str>,
-    font_style: u8,
-}
-
-#[allow(clippy::too_many_arguments)]
-fn push_with_capture_ranges(
-    out: &mut Vec<StyledJsonToken>,
-    utf16_map: &[usize],
-    code: &str,
-    match_start_utf16: usize,
-    match_end_utf16: usize,
-    base_color: &Arc<str>,
-    base_font_style: u8,
-    mut capture_ranges: Vec<CaptureRange>,
-) -> Result<()> {
-    if capture_ranges.is_empty() {
-        return push_styled_slice(
-            out,
-            utf16_map,
-            code,
-            match_start_utf16,
-            match_end_utf16,
-            base_color,
-            base_font_style,
-        );
-    }
-
-    capture_ranges.sort_by(|a, b| a.start.cmp(&b.start).then(a.end.cmp(&b.end)));
-
-    let mut cursor = match_start_utf16;
-    for cr in &capture_ranges {
-        if cr.start > cursor {
-            push_styled_slice(
-                out,
-                utf16_map,
-                code,
-                cursor,
-                cr.start,
-                base_color,
-                base_font_style,
-            )?;
-        }
-        let segment_start = cr.start.max(cursor);
-        if cr.end > segment_start {
-            push_styled_slice(
-                out,
-                utf16_map,
-                code,
-                segment_start,
-                cr.end,
-                &cr.color,
-                cr.font_style,
-            )?;
-            cursor = cr.end;
-        }
-    }
-
-    if match_end_utf16 > cursor {
-        push_styled_slice(
-            out,
-            utf16_map,
-            code,
-            cursor,
-            match_end_utf16,
-            base_color,
-            base_font_style,
-        )?;
-    }
-
-    Ok(())
-}
-
-fn escape_regex_literal(input: &str) -> String {
-    let mut out = String::with_capacity(input.len());
-    for ch in input.chars() {
-        match ch {
-            '\\' | '.' | '+' | '*' | '?' | '(' | ')' | '[' | ']' | '{' | '}' | '^' | '$' | '|' => {
-                out.push('\\');
-                out.push(ch);
-            }
-            _ => out.push(ch),
-        }
-    }
-    out
-}
-
-fn resolve_pattern_backrefs(
-    pattern: &str,
-    capture_indices: &[ferroni::scanner::CaptureIndex],
-    utf16_map: &[usize],
-    code: &str,
-) -> String {
-    let mut out = String::with_capacity(pattern.len());
-    let mut chars = pattern.chars().peekable();
-
-    while let Some(ch) = chars.next() {
-        if ch != '\\' {
-            out.push(ch);
-            continue;
-        }
-
-        let mut digits = String::new();
-        while let Some(next) = chars.peek() {
-            if next.is_ascii_digit() {
-                digits.push(*next);
-                chars.next();
-            } else {
-                break;
-            }
-        }
-
-        if digits.is_empty() {
-            out.push('\\');
-            continue;
-        }
-
-        let Ok(index) = digits.parse::<usize>() else {
-            continue;
-        };
-        let Some(range) = capture_indices.get(index) else {
-            continue;
-        };
-        if range.end <= range.start {
-            continue;
-        }
-        if range.start >= utf16_map.len() || range.end >= utf16_map.len() {
-            continue;
-        }
-
-        let start_byte = utf16_map[range.start];
-        let end_byte = utf16_map[range.end];
-        let Some(captured) = code.get(start_byte..end_byte) else {
-            continue;
-        };
-
-        out.push_str(&escape_regex_literal(captured));
-    }
-
-    out
-}
-
-fn resolve_capture_name_backrefs(
-    pattern: &str,
-    capture_indices: &[ferroni::scanner::CaptureIndex],
-    utf16_map: &[usize],
-    code: &str,
-) -> String {
-    let mut out = String::with_capacity(pattern.len());
-    let chars = pattern.as_bytes();
-    let mut index = 0usize;
-
-    while index < chars.len() {
-        if chars[index] != b'$' {
-            out.push(chars[index] as char);
-            index += 1;
-            continue;
-        }
-
-        if index + 1 >= chars.len() {
-            out.push('$');
-            index += 1;
-            continue;
-        }
-
-        if chars[index + 1] == b'{' {
-            let mut cursor = index + 2;
-            let mut digits = String::new();
-            while cursor < chars.len() && chars[cursor].is_ascii_digit() {
-                digits.push(chars[cursor] as char);
-                cursor += 1;
-            }
-
-            let mut transform = None;
-            if cursor + 2 < chars.len() && chars[cursor] == b':' && chars[cursor + 1] == b'/' {
-                cursor += 2;
-                let start = cursor;
-                while cursor < chars.len() && chars[cursor] != b'}' {
-                    cursor += 1;
-                }
-                if cursor <= chars.len() {
-                    transform = std::str::from_utf8(&chars[start..cursor]).ok();
-                }
-            }
-
-            if cursor < chars.len() && chars[cursor] == b'}' {
-                if let Some(captured) =
-                    resolve_capture_reference(&digits, capture_indices, utf16_map, code)
-                {
-                    match transform {
-                        Some("downcase") => out.push_str(&captured.to_lowercase()),
-                        Some("upcase") => out.push_str(&captured.to_uppercase()),
-                        _ => out.push_str(&captured),
-                    }
-                    index = cursor + 1;
-                    continue;
-                }
-            }
-        }
-
-        let mut cursor = index + 1;
-        let mut digits = String::new();
-        while cursor < chars.len() && chars[cursor].is_ascii_digit() {
-            digits.push(chars[cursor] as char);
-            cursor += 1;
-        }
-        if let Some(captured) = resolve_capture_reference(&digits, capture_indices, utf16_map, code)
-        {
-            out.push_str(&captured);
-            index = cursor;
-            continue;
-        }
-
-        out.push('$');
-        index += 1;
-    }
-
-    out
-}
-
-fn resolve_capture_reference(
-    digits: &str,
-    capture_indices: &[ferroni::scanner::CaptureIndex],
-    utf16_map: &[usize],
-    code: &str,
-) -> Option<String> {
-    let index = digits.parse::<usize>().ok()?;
-    let range = capture_indices.get(index)?;
-    if range.end <= range.start || range.start >= utf16_map.len() || range.end >= utf16_map.len() {
-        return None;
-    }
-    let start_byte = utf16_map[range.start];
-    let end_byte = utf16_map[range.end];
-    code.get(start_byte..end_byte).map(str::to_owned)
-}
-
-fn build_scope_stack_from_frames(stack: &[StateFrame], root_scope: Option<&str>) -> Vec<String> {
-    let mut scopes = Vec::new();
-    if let Some(root) = root_scope {
-        scopes.push(root.to_owned());
-    }
-    for frame in stack {
-        for s in &frame.name_scopes {
-            scopes.push(s.clone());
-        }
-        for s in &frame.content_scopes {
-            scopes.push(s.clone());
-        }
-    }
-    scopes
-}
-
-fn resolve_color_for_scope_stack_owned(
-    scope_stack: &[String],
-    theme: &ThemeData,
-    cache: &mut ThemeCache,
-) -> (Arc<str>, u8) {
-    let style = cache.resolve_owned(scope_stack, theme);
-    let color = style
-        .foreground
-        .clone()
-        .unwrap_or_else(|| theme.fg_normalized.clone());
-    (color, style.font_style)
-}
-
-fn resolve_color_with_extra_scope(
-    scope_stack: &[String],
-    extra: &str,
-    theme: &ThemeData,
-    cache: &mut ThemeCache,
-) -> (Arc<str>, u8) {
-    let style = cache.resolve_with_extra_owned(scope_stack, extra, theme);
-    let color = style
-        .foreground
-        .clone()
-        .unwrap_or_else(|| theme.fg_normalized.clone());
-    (color, style.font_style)
-}
-
-// ─────────────────────────────────────────────────────────────────────────────
-// Main tokenization loop (port of vscode-textmate _tokenizeString)
-// ─────────────────────────────────────────────────────────────────────────────
-
-fn tokenize_with_grammar_skeleton(
-    code: &str,
-    compiled: &mut CompiledGrammar,
-    root_scope: Option<&str>,
-    theme: &ThemeData,
-    initial_stack: Option<Vec<StateFrame>>,
-) -> Result<(Vec<StyledJsonToken>, Vec<StateFrame>)> {
-    let root_rule_id = compiled.root_rule_id;
-
-    if !compiled.scanner_cache.contains_key(&(root_rule_id, None)) {
-        let default_fg = theme.fg_normalized.clone();
-        let default_stack = initial_stack.unwrap_or_else(|| {
-            vec![StateFrame {
-                rule_id: root_rule_id,
-                _enter_pos: -1,
-                _anchor_pos: 0,
-                end_rule: None,
-                name_scopes: Vec::new(),
-                content_scopes: Vec::new(),
-            }]
-        });
-        let utf16_len = code.encode_utf16().count();
-        return Ok((
-            vec![StyledJsonToken {
-                content: code.to_owned(),
-                content_utf16_len: utf16_len,
-                offset_utf16: 0,
-                color: default_fg,
-                font_style: 0,
-                dark_color: None,
-            }],
-            default_stack,
-        ));
-    }
-
-    // Build state stack
-    let mut stack: Vec<StateFrame> = initial_stack.unwrap_or_else(|| {
-        vec![StateFrame {
-            rule_id: root_rule_id,
-            _enter_pos: -1,
-            _anchor_pos: 0,
-            end_rule: None,
-            name_scopes: Vec::new(),
-            content_scopes: Vec::new(),
-        }]
-    });
-
-    // ── Global UTF-16 map (for output positioning) ──
-    let utf16_map = utf16_to_byte_map(code);
-    let total_utf16 = utf16_map.len().saturating_sub(1);
-    let find_options = ScannerFindOptions::from_bits(0);
-    let mut out = Vec::new();
-
-    // Safeguards against infinite loops
-    let max_iterations = code.len().saturating_mul(10).max(50_000);
-    let mut iterations = 0usize;
-    let max_stack_depth: usize = 64;
-    let deadline = Instant::now() + Duration::from_secs(30);
-
-    // ── Theme resolution cache ──
-    let mut theme_cache = ThemeCache::new();
-
-    // ── Scope-stack & color cache ──
-    let mut stack_generation: u64 = 0;
-    let mut cached_scope_stack: Vec<String> = build_scope_stack_from_frames(&stack, root_scope);
-    let (mut cached_color, mut cached_font_style) =
-        resolve_color_for_scope_stack_owned(&cached_scope_stack, theme, &mut theme_cache);
-    let mut cached_generation: u64 = 0;
-
-    // ── Per-line tokenization (like vscode-textmate) ──
-    // Split code into lines and process each line with a line-local OnigString.
-    // This avoids O(n²) scanning of the full code string on every match.
-    let lines: Vec<&str> = code.split('\n').collect();
-    let mut global_offset_utf16: usize = 0;
-    // Cache frame info to avoid repeated extraction when stack hasn't changed
-    let mut last_cache_key: Option<(RuleId, Option<String>)> = None;
-    let mut cached_frame_is_while: bool = false;
-    let mut cached_frame_apply_end_last: bool = false;
-    let mut injection_cache_generation: u64 = u64::MAX;
-    let mut active_injections: Vec<(RuleId, InjectionPriority)> = Vec::new();
-    let mut selector_scope_match_cache: Vec<HashMap<String, bool>> = (0..compiled.injections.len())
-        .map(|_| HashMap::new())
-        .collect();
-
-    'line_loop: for (line_idx, &line_text) in lines.iter().enumerate() {
-        // Build line text with trailing \n (except last line)
-        let has_newline = line_idx < lines.len() - 1;
-        let line_with_nl: String;
-        let line_str: &str = if has_newline {
-            let mut buf = String::with_capacity(line_text.len() + 1);
-            buf.push_str(line_text);
-            buf.push('\n');
-            line_with_nl = buf;
-            &line_with_nl
-        } else {
-            line_text
-        };
-        let line_input = OnigString::new(line_str);
-        let line_str_id = NEXT_ONIG_STR_ID.fetch_add(1, Ordering::Relaxed);
-        let line_utf16_map = utf16_to_byte_map(line_str);
-        let line_utf16_len = line_utf16_map.len().saturating_sub(1);
-
-        if line_utf16_len == 0 {
-            global_offset_utf16 += if has_newline { 1 } else { 0 };
-            continue;
-        }
-
-        let mut cursor: usize = 0; // line-local cursor (UTF-16 units)
-        let mut zero_width_count = 0usize;
-        let mut last_zero_width_pos = usize::MAX;
-        let mut last_zero_width_generation = u64::MAX;
-
-        // While-condition check at start of each line (except first)
-        if line_idx > 0 && !stack.is_empty() {
-            let top_rule_id = stack.last().map(|f| f.rule_id).unwrap_or(0);
-            let top_is_while = matches!(
-                compiled.registry.get(top_rule_id),
-                Some(Rule::BeginWhile { .. })
-            );
-            if top_is_while {
-                let while_re = {
-                    let frame = stack.last().unwrap();
-                    if let Some(end_rule) = &frame.end_rule {
-                        Some(end_rule.clone())
-                    } else if let Some(Rule::BeginWhile { while_re, .. }) =
-                        compiled.registry.get(top_rule_id)
-                    {
-                        Some(while_re.clone())
-                    } else {
-                        None
-                    }
-                };
-
-                if let Some(while_re) = while_re {
-                    let while_matched = {
-                        let scanner = match compiled.while_scanner_cache.entry(while_re.clone()) {
-                            std::collections::hash_map::Entry::Occupied(entry) => entry.into_mut(),
-                            std::collections::hash_map::Entry::Vacant(entry) => {
-                                let scanner =
-                                    Scanner::new(&[while_re.as_str()]).map_err(|err| {
-                                        Error::from_reason(format!(
-                                            "Failed to compile while pattern: {err}"
-                                        ))
-                                    })?;
-                                entry.insert(scanner)
-                            }
-                        };
-                        scanner.find_next_match_utf16(&line_input, 0, find_options)
-                    };
-
-                    if cached_generation != stack_generation {
-                        cached_scope_stack = build_scope_stack_from_frames(&stack, root_scope);
-                        let (c, fs) = resolve_color_for_scope_stack_owned(
-                            &cached_scope_stack,
-                            theme,
-                            &mut theme_cache,
-                        );
-                        cached_color = c;
-                        cached_font_style = fs;
-                        cached_generation = stack_generation;
-                    }
-
-                    if let Some(found_while) = while_matched {
-                        let first = found_while.capture_indices.first().ok_or_else(|| {
-                            Error::from_reason("While scanner returned match without capture 0.")
-                        })?;
-                        let while_start = first.start;
-                        let while_end = first.end;
-
-                        if while_start != 0 {
-                            if stack.len() > 1 {
-                                stack.pop();
-                                stack_generation += 1;
-                                // Don't skip line — continue tokenizing it with the new stack
-                            }
-                        } else {
-                            // While matched — handle captures and advance cursor
-                            let while_captures: &[GrammarCapture] =
-                                if let Some(Rule::BeginWhile { while_captures, .. }) =
-                                    compiled.registry.get(top_rule_id)
-                                {
-                                    while_captures
-                                } else {
-                                    &[]
-                                };
-
-                            let scope_stack = &cached_scope_stack;
-                            let mut capture_ranges = Vec::new();
-                            for capture in while_captures {
-                                let Some(range) = found_while.capture_indices.get(capture.index)
-                                else {
-                                    continue;
-                                };
-                                if range.end <= range.start {
-                                    continue;
-                                }
-                                if range.start < while_start || range.end > while_end {
-                                    continue;
-                                }
-                                let (cap_color, cap_fs) =
-                                    if let Some(name) = capture.name.as_deref() {
-                                        let resolved_name = resolve_capture_name_backrefs(
-                                            name,
-                                            &found_while.capture_indices,
-                                            &line_utf16_map,
-                                            line_str,
-                                        );
-                                        resolve_color_with_extra_scope(
-                                            scope_stack,
-                                            &resolved_name,
-                                            theme,
-                                            &mut theme_cache,
-                                        )
-                                    } else {
-                                        (cached_color.clone(), cached_font_style)
-                                    };
-                                capture_ranges.push(CaptureRange {
-                                    start: while_start + global_offset_utf16,
-                                    end: range.end + global_offset_utf16,
-                                    color: cap_color,
-                                    font_style: cap_fs,
-                                });
-                            }
-                            // Adjust capture ranges back to global coordinates for output
-                            let global_while_start = while_start + global_offset_utf16;
-                            let global_while_end = while_end + global_offset_utf16;
-                            let mut cap_ranges_global = Vec::new();
-                            for capture in while_captures {
-                                let Some(range) = found_while.capture_indices.get(capture.index)
-                                else {
-                                    continue;
-                                };
-                                if range.end <= range.start {
-                                    continue;
-                                }
-                                if range.start < while_start || range.end > while_end {
-                                    continue;
-                                }
-                                let (cap_color, cap_fs) =
-                                    if let Some(name) = capture.name.as_deref() {
-                                        let resolved_name = resolve_capture_name_backrefs(
-                                            name,
-                                            &found_while.capture_indices,
-                                            &line_utf16_map,
-                                            line_str,
-                                        );
-                                        resolve_color_with_extra_scope(
-                                            &cached_scope_stack,
-                                            &resolved_name,
-                                            theme,
-                                            &mut theme_cache,
-                                        )
-                                    } else {
-                                        (cached_color.clone(), cached_font_style)
-                                    };
-                                cap_ranges_global.push(CaptureRange {
-                                    start: range.start + global_offset_utf16,
-                                    end: range.end + global_offset_utf16,
-                                    color: cap_color,
-                                    font_style: cap_fs,
-                                });
-                            }
-                            if global_while_end > global_while_start {
-                                push_with_capture_ranges(
-                                    &mut out,
-                                    &utf16_map,
-                                    code,
-                                    global_while_start,
-                                    global_while_end,
-                                    &cached_color,
-                                    cached_font_style,
-                                    cap_ranges_global,
-                                )?;
-                            }
-                            cursor = while_end;
-                        }
-                    } else if stack.len() > 1 {
-                        stack.pop();
-                        stack_generation += 1;
-                    }
-                }
-            }
-        }
-
-        // ── Inner loop: process matches within this line ──
-        while cursor < line_utf16_len {
-            iterations += 1;
-            if iterations > max_iterations || (iterations & 1023 == 0 && Instant::now() > deadline)
-            {
-                // Bail out — emit rest of file as single token
-                let remaining_global = cursor + global_offset_utf16;
-                if remaining_global < total_utf16 {
-                    if cached_generation != stack_generation {
-                        cached_scope_stack = build_scope_stack_from_frames(&stack, root_scope);
-                        let (c, fs) = resolve_color_for_scope_stack_owned(
-                            &cached_scope_stack,
-                            theme,
-                            &mut theme_cache,
-                        );
-                        cached_color = c;
-                        cached_font_style = fs;
-                    }
-                    push_styled_slice(
-                        &mut out,
-                        &utf16_map,
-                        code,
-                        remaining_global,
-                        total_utf16,
-                        &cached_color,
-                        cached_font_style,
-                    )?;
-                }
-                break 'line_loop;
-            }
-
-            if stack.is_empty() {
-                break 'line_loop;
-            }
-
-            // Get or build scanner for current frame (cached across iterations)
-            if cached_generation != stack_generation || last_cache_key.is_none() {
-                let frame_rule_id_new = stack.last().map(|f| f.rule_id).unwrap_or(0);
-                let frame_is_while_new = matches!(
-                    compiled.registry.get(frame_rule_id_new),
-                    Some(Rule::BeginWhile { .. })
-                );
-                let frame_end_rule_new = if frame_is_while_new {
-                    None
-                } else {
-                    stack.last().and_then(|f| f.end_rule.clone())
-                };
-                let frame_apply_end_last_new = if let Some(Rule::BeginEnd {
-                    apply_end_pattern_last,
-                    ..
-                }) = compiled.registry.get(frame_rule_id_new)
-                {
-                    *apply_end_pattern_last
-                } else {
-                    false
-                };
-                last_cache_key = Some((frame_rule_id_new, frame_end_rule_new));
-                cached_frame_is_while = frame_is_while_new;
-                cached_frame_apply_end_last = frame_apply_end_last_new;
-            }
-            let cache_key = last_cache_key.as_ref().unwrap();
-            let frame_rule_id = cache_key.0;
-            let frame_is_while = cached_frame_is_while;
-            if !compiled.scanner_cache.contains_key(cache_key) {
-                let scanner = build_scanner_for_rule(
-                    frame_rule_id,
-                    &compiled.registry,
-                    cache_key.1.as_deref(),
-                    cached_frame_apply_end_last,
-                );
-                if let Some(scanner) = scanner {
-                    compiled.scanner_cache.insert(cache_key.clone(), scanner);
-                }
-            }
-
-            // Refresh scope stack + color cache if stack changed
-            if cached_generation != stack_generation {
-                cached_scope_stack = build_scope_stack_from_frames(&stack, root_scope);
-                let (c, fs) = resolve_color_for_scope_stack_owned(
-                    &cached_scope_stack,
-                    theme,
-                    &mut theme_cache,
-                );
-                cached_color = c;
-                cached_font_style = fs;
-                cached_generation = stack_generation;
-            }
-            let scope_stack = &cached_scope_stack;
-            let inherited_color = &cached_color;
-            let inherited_font_style = cached_font_style;
-
-            if !compiled.injections.is_empty() && injection_cache_generation != stack_generation {
-                active_injections.clear();
-                for (idx, injection) in compiled.injections.iter().enumerate() {
-                    let scope_cache = &mut selector_scope_match_cache[idx];
-                    let scope_key = scope_stack.join("\u{1f}");
-                    let matches_scope = if let Some(cached) = scope_cache.get(&scope_key) {
-                        *cached
-                    } else {
-                        let parsed =
-                            selector_matches_compiled(&injection.compiled_selector, scope_stack);
-                        scope_cache.insert(scope_key, parsed);
-                        parsed
-                    };
-
-                    if matches_scope {
-                        active_injections.push((injection.rule_id, injection.priority));
-                    }
-                }
-                injection_cache_generation = stack_generation;
-            }
-
-            // Scanner uses line-local OnigString and cursor
-            let main_match = if let Some(cached_scanner) = compiled.scanner_cache.get_mut(cache_key)
-            {
-                let m = find_next_match_ordered(
-                    cached_scanner,
-                    &line_input,
-                    line_str_id,
-                    cursor,
-                    find_options,
-                );
-                m.map(|m| {
-                    let rule_id = cached_scanner.rule_ids.get(m.index).copied().unwrap_or(0);
-                    (m, rule_id)
-                })
-            } else {
-                None
-            };
-
-            // Injection matches also use line-local input
-            let injection_match = if !active_injections.is_empty() {
-                find_injection_match(
-                    compiled,
-                    &active_injections,
-                    &line_input,
-                    line_str_id,
-                    cursor,
-                    find_options,
-                )
-            } else {
-                None
-            };
-
-            // Pick the best match
-            let (found, matched_rule_id) = match (main_match, injection_match) {
-                (None, None) => {
-                    // No match in this line — emit rest of line
-                    if frame_is_while {
-                        // For while frames, emit to end of line (while check happens at next line start)
-                        let g_cursor = cursor + global_offset_utf16;
-                        let g_end = line_utf16_len + global_offset_utf16;
-                        push_styled_slice(
-                            &mut out,
-                            &utf16_map,
-                            code,
-                            g_cursor,
-                            g_end,
-                            inherited_color,
-                            inherited_font_style,
-                        )?;
-                    } else {
-                        let g_cursor = cursor + global_offset_utf16;
-                        let g_end = line_utf16_len + global_offset_utf16;
-                        push_styled_slice(
-                            &mut out,
-                            &utf16_map,
-                            code,
-                            g_cursor,
-                            g_end,
-                            inherited_color,
-                            inherited_font_style,
-                        )?;
-                    }
-                    break; // next line
-                }
-                (Some(m), None) => m,
-                (None, Some((inj_found, inj_id, _))) => (inj_found, inj_id),
-                (Some((main_found, main_id)), Some((inj_found, inj_id, inj_prio))) => {
-                    let main_start = main_found
-                        .capture_indices
-                        .first()
-                        .map(|c| c.start)
-                        .unwrap_or(usize::MAX);
-                    let inj_start = inj_found
-                        .capture_indices
-                        .first()
-                        .map(|c| c.start)
-                        .unwrap_or(usize::MAX);
-                    if inj_start < main_start
-                        || (inj_start == main_start && inj_prio == InjectionPriority::Left)
-                    {
-                        (inj_found, inj_id)
-                    } else {
-                        (main_found, main_id)
-                    }
-                }
-            };
-
-            let first = found.capture_indices.first().ok_or_else(|| {
-                Error::from_reason("Ferriki grammar scanner returned match without capture 0.")
-            })?;
-            // Line-local match positions
-            let start_local = first.start;
-            let end_local = first.end;
-            // Global positions for output
-            let start_utf16 = start_local + global_offset_utf16;
-            let end_utf16 = end_local + global_offset_utf16;
-
-            if start_local > cursor {
-                let g_cursor = cursor + global_offset_utf16;
-                push_styled_slice(
-                    &mut out,
-                    &utf16_map,
-                    code,
-                    g_cursor,
-                    start_utf16,
-                    inherited_color,
-                    inherited_font_style,
-                )?;
-                cursor = start_local;
-                continue;
-            }
-
-            // Zero-width handling
-            if end_local <= start_local {
-                match matched_rule_id {
-                    END_RULE_ID => {
-                        if stack.len() > 1 {
-                            stack.pop();
-                            stack_generation += 1;
-                        }
-                        continue;
-                    }
-                    _ => {
-                        let is_begin = matches!(
-                            compiled.registry.get(matched_rule_id),
-                            Some(Rule::BeginEnd { .. }) | Some(Rule::BeginWhile { .. })
-                        );
-                        if !is_begin {
-                            cursor = start_local.saturating_add(1);
-                            continue;
-                        }
-                        // Fall through to begin handling below
-                    }
-                }
-            }
-
-            // Dispatch based on matched rule type
-            if matched_rule_id == END_RULE_ID {
-                let end_captures: &[GrammarCapture] =
-                    if let Some(Rule::BeginEnd { end_captures, .. }) =
-                        compiled.registry.get(frame_rule_id)
-                    {
-                        end_captures
-                    } else {
-                        &[]
-                    };
-
-                let mut capture_ranges = Vec::new();
-                for capture in end_captures {
-                    let Some(range) = found.capture_indices.get(capture.index) else {
-                        continue;
-                    };
-                    if range.end <= range.start {
-                        continue;
-                    }
-                    if range.start < start_local || range.end > end_local {
-                        continue;
-                    }
-                    let (cap_color, cap_fs) = if let Some(name) = capture.name.as_deref() {
-                        let resolved_name = resolve_capture_name_backrefs(
-                            name,
-                            &found.capture_indices,
-                            &line_utf16_map,
-                            line_str,
-                        );
-                        resolve_color_with_extra_scope(
-                            scope_stack,
-                            &resolved_name,
-                            theme,
-                            &mut theme_cache,
-                        )
-                    } else {
-                        (inherited_color.clone(), inherited_font_style)
-                    };
-                    capture_ranges.push(CaptureRange {
-                        start: range.start + global_offset_utf16,
-                        end: range.end + global_offset_utf16,
-                        color: cap_color,
-                        font_style: cap_fs,
-                    });
-                }
-                if capture_ranges.is_empty() && end_captures.len() == 1 && end_utf16 > start_utf16 {
-                    if let Some(name) = end_captures[0].name.as_deref() {
-                        let resolved_name = resolve_capture_name_backrefs(
-                            name,
-                            &found.capture_indices,
-                            &line_utf16_map,
-                            line_str,
-                        );
-                        let (cap_color, cap_fs) = resolve_color_with_extra_scope(
-                            scope_stack,
-                            &resolved_name,
-                            theme,
-                            &mut theme_cache,
-                        );
-                        capture_ranges.push(CaptureRange {
-                            start: start_utf16,
-                            end: end_utf16,
-                            color: cap_color,
-                            font_style: cap_fs,
-                        });
-                    }
-                }
-                push_with_capture_ranges(
-                    &mut out,
-                    &utf16_map,
-                    code,
-                    start_utf16,
-                    end_utf16,
-                    inherited_color,
-                    inherited_font_style,
-                    capture_ranges,
-                )?;
-                if stack.len() > 1 {
-                    stack.pop();
-                    stack_generation += 1;
-                }
-            } else {
-                let matched_rule = compiled.registry.get(matched_rule_id);
-
-                match matched_rule {
-                    Some(Rule::Match { name, captures, .. }) => {
-                        let (color, font_style) = if let Some(n) = name.as_deref() {
-                            resolve_color_with_extra_scope(scope_stack, n, theme, &mut theme_cache)
-                        } else {
-                            (inherited_color.clone(), inherited_font_style)
-                        };
-                        let mut capture_ranges = Vec::new();
-                        for capture in captures {
-                            let Some(range) = found.capture_indices.get(capture.index) else {
-                                continue;
-                            };
-                            if range.end <= range.start {
-                                continue;
-                            }
-                            if range.start < start_local || range.end > end_local {
-                                continue;
-                            }
-                            let (cap_color, cap_fs) = if let Some(n) = capture.name.as_deref() {
-                                let resolved_name = resolve_capture_name_backrefs(
-                                    n,
-                                    &found.capture_indices,
-                                    &line_utf16_map,
-                                    line_str,
-                                );
-                                resolve_color_with_extra_scope(
-                                    scope_stack,
-                                    &resolved_name,
-                                    theme,
-                                    &mut theme_cache,
-                                )
-                            } else {
-                                (color.clone(), font_style)
-                            };
-                            capture_ranges.push(CaptureRange {
-                                start: range.start + global_offset_utf16,
-                                end: range.end + global_offset_utf16,
-                                color: cap_color,
-                                font_style: cap_fs,
-                            });
-                        }
-                        push_with_capture_ranges(
-                            &mut out,
-                            &utf16_map,
-                            code,
-                            start_utf16,
-                            end_utf16,
-                            &color,
-                            font_style,
-                            capture_ranges,
-                        )?;
-                    }
-
-                    Some(Rule::BeginEnd {
-                        name,
-                        content_name,
-                        end_re,
-                        end_has_back_references,
-                        begin_captures,
-                        ..
-                    }) => {
-                        let resolved_end_re = if *end_has_back_references {
-                            resolve_pattern_backrefs(
-                                end_re,
-                                &found.capture_indices,
-                                &line_utf16_map,
-                                line_str,
-                            )
-                        } else {
-                            end_re.clone()
-                        };
-
-                        let (color, font_style) = if let Some(n) = name.as_deref() {
-                            resolve_color_with_extra_scope(scope_stack, n, theme, &mut theme_cache)
-                        } else {
-                            (inherited_color.clone(), inherited_font_style)
-                        };
-
-                        let mut capture_ranges = Vec::new();
-                        for capture in begin_captures {
-                            let Some(range) = found.capture_indices.get(capture.index) else {
-                                continue;
-                            };
-                            if range.end <= range.start {
-                                continue;
-                            }
-                            if range.start < start_local || range.end > end_local {
-                                continue;
-                            }
-                            let (cap_color, cap_fs) = if let Some(n) = capture.name.as_deref() {
-                                let resolved_name = resolve_capture_name_backrefs(
-                                    n,
-                                    &found.capture_indices,
-                                    &line_utf16_map,
-                                    line_str,
-                                );
-                                resolve_color_with_extra_scope(
-                                    scope_stack,
-                                    &resolved_name,
-                                    theme,
-                                    &mut theme_cache,
-                                )
-                            } else {
-                                (color.clone(), font_style)
-                            };
-                            capture_ranges.push(CaptureRange {
-                                start: range.start + global_offset_utf16,
-                                end: range.end + global_offset_utf16,
-                                color: cap_color,
-                                font_style: cap_fs,
-                            });
-                        }
-
-                        if end_utf16 > start_utf16 {
-                            push_with_capture_ranges(
-                                &mut out,
-                                &utf16_map,
-                                code,
-                                start_utf16,
-                                end_utf16,
-                                &color,
-                                font_style,
-                                capture_ranges,
-                            )?;
-                        }
-
-                        let name_scopes = name.as_deref().map(parse_scope_list).unwrap_or_default();
-                        let content_scopes = content_name
-                            .as_deref()
-                            .map(parse_scope_list)
-                            .unwrap_or_default();
-
-                        stack.push(StateFrame {
-                            rule_id: matched_rule_id,
-                            _enter_pos: start_utf16 as i32,
-                            _anchor_pos: end_utf16 as i32,
-                            end_rule: Some(resolved_end_re),
-                            name_scopes,
-                            content_scopes,
-                        });
-                        stack_generation += 1;
-                    }
-
-                    Some(Rule::BeginWhile {
-                        name,
-                        content_name,
-                        while_re,
-                        while_has_back_references,
-                        begin_captures,
-                        ..
-                    }) => {
-                        let resolved_while_re = if *while_has_back_references {
-                            resolve_pattern_backrefs(
-                                while_re,
-                                &found.capture_indices,
-                                &line_utf16_map,
-                                line_str,
-                            )
-                        } else {
-                            while_re.clone()
-                        };
-
-                        let (color, font_style) = if let Some(n) = name.as_deref() {
-                            resolve_color_with_extra_scope(scope_stack, n, theme, &mut theme_cache)
-                        } else {
-                            (inherited_color.clone(), inherited_font_style)
-                        };
-
-                        let mut capture_ranges = Vec::new();
-                        for capture in begin_captures {
-                            let Some(range) = found.capture_indices.get(capture.index) else {
-                                continue;
-                            };
-                            if range.end <= range.start {
-                                continue;
-                            }
-                            if range.start < start_local || range.end > end_local {
-                                continue;
-                            }
-                            let (cap_color, cap_fs) = if let Some(n) = capture.name.as_deref() {
-                                let resolved_name = resolve_capture_name_backrefs(
-                                    n,
-                                    &found.capture_indices,
-                                    &line_utf16_map,
-                                    line_str,
-                                );
-                                resolve_color_with_extra_scope(
-                                    scope_stack,
-                                    &resolved_name,
-                                    theme,
-                                    &mut theme_cache,
-                                )
-                            } else {
-                                (color.clone(), font_style)
-                            };
-                            capture_ranges.push(CaptureRange {
-                                start: range.start + global_offset_utf16,
-                                end: range.end + global_offset_utf16,
-                                color: cap_color,
-                                font_style: cap_fs,
-                            });
-                        }
-
-                        if end_utf16 > start_utf16 {
-                            push_with_capture_ranges(
-                                &mut out,
-                                &utf16_map,
-                                code,
-                                start_utf16,
-                                end_utf16,
-                                &color,
-                                font_style,
-                                capture_ranges,
-                            )?;
-                        }
-
-                        let name_scopes = name.as_deref().map(parse_scope_list).unwrap_or_default();
-                        let content_scopes = content_name
-                            .as_deref()
-                            .map(parse_scope_list)
-                            .unwrap_or_default();
-
-                        stack.push(StateFrame {
-                            rule_id: matched_rule_id,
-                            _enter_pos: start_utf16 as i32,
-                            _anchor_pos: end_utf16 as i32,
-                            end_rule: Some(resolved_while_re),
-                            name_scopes,
-                            content_scopes,
-                        });
-                        stack_generation += 1;
-                    }
-
-                    Some(Rule::IncludeOnly { .. }) | None => {
-                        cursor = end_local.max(cursor.saturating_add(1));
-                        continue;
-                    }
-                }
-            }
-
-            // Zero-width loop detection
-            if end_local == cursor {
-                if cursor == last_zero_width_pos && stack_generation == last_zero_width_generation {
-                    zero_width_count += 1;
-                    if zero_width_count > 3 {
-                        cursor = cursor.saturating_add(1);
-                        zero_width_count = 0;
-                        continue;
-                    }
-                } else {
-                    last_zero_width_pos = cursor;
-                    last_zero_width_generation = stack_generation;
-                    zero_width_count = 1;
-                }
-            } else {
-                zero_width_count = 0;
-            }
-
-            // Stack depth limit
-            if stack.len() > max_stack_depth {
-                if cached_generation != stack_generation {
-                    cached_scope_stack = build_scope_stack_from_frames(&stack, root_scope);
-                    let (c, fs) = resolve_color_for_scope_stack_owned(
-                        &cached_scope_stack,
-                        theme,
-                        &mut theme_cache,
-                    );
-                    cached_color = c;
-                    cached_font_style = fs;
-                }
-                push_styled_slice(
-                    &mut out,
-                    &utf16_map,
-                    code,
-                    end_utf16,
-                    total_utf16,
-                    &cached_color,
-                    cached_font_style,
-                )?;
-                break 'line_loop;
-            }
-
-            cursor = end_local;
-        }
-
-        // Advance global offset past this line.
-        // line_utf16_len already includes the \n when has_newline is true,
-        // because line_str includes it.
-        global_offset_utf16 += line_utf16_len;
-    }
-
-    // Merge adjacent tokens with the same color and font_style
-    let mut merged = Vec::with_capacity(out.len());
-    for token in out {
-        if let Some(last) = merged.last_mut() {
-            let last: &mut StyledJsonToken = last;
-            let last_end_utf16 = last.offset_utf16 + last.content_utf16_len;
-            if last.color == token.color
-                && last.font_style == token.font_style
-                && last_end_utf16 == token.offset_utf16
-            {
-                last.content.push_str(&token.content);
-                last.content_utf16_len += token.content_utf16_len;
-                continue;
-            }
-        }
-        merged.push(token);
-    }
-
-    Ok((merged, stack))
-}
-
-/// Try to find the best injection match at the current position.
-/// Returns (match, matched_rule_id, priority) — extracting the single matched rule_id
-/// instead of cloning the entire rule_ids Vec.
-fn find_injection_match(
-    compiled: &mut CompiledGrammar,
-    active_injections: &[(RuleId, InjectionPriority)],
-    input: &OnigString,
-    line_str_id: u64,
-    cursor: usize,
-    find_options: ScannerFindOptions,
-) -> Option<(ferroni::scanner::ScannerMatch, RuleId, InjectionPriority)> {
-    let mut best_result: Option<(
-        ferroni::scanner::ScannerMatch,
-        RuleId,
-        usize,
-        InjectionPriority,
-    )> = None;
-
-    for (rule_id, priority) in active_injections {
-        // Build and cache injection scanner
-        if !compiled.injection_scanner_cache.contains_key(rule_id) {
-            let mut pattern_pairs: Vec<(String, RuleId)> = Vec::new();
-            collect_patterns(*rule_id, &compiled.registry, &mut pattern_pairs);
-            if pattern_pairs.is_empty() {
-                continue;
-            }
-
-            let regexes: Vec<String> = pattern_pairs.iter().map(|(re, _)| re.clone()).collect();
-            let ids: Vec<RuleId> = pattern_pairs.iter().map(|(_, id)| *id).collect();
-            let regex_refs: Vec<&str> = regexes.iter().map(String::as_str).collect();
-            let Ok(scanner) = Scanner::new(&regex_refs) else {
-                continue;
-            };
-            let single_scanners = std::iter::repeat_with(|| None)
-                .take(regexes.len())
-                .collect();
-            compiled.injection_scanner_cache.insert(
-                *rule_id,
-                CompiledScanner {
-                    scanner,
-                    rule_ids: ids,
-                    regexes,
-                    single_scanners,
-                },
-            );
-        }
-
-        let Some(cached) = compiled.injection_scanner_cache.get_mut(rule_id) else {
-            continue;
-        };
-        let Some(found) = find_next_match_ordered(cached, input, line_str_id, cursor, find_options)
-        else {
-            continue;
-        };
-
-        let start = found
-            .capture_indices
-            .first()
-            .map(|c| c.start)
-            .unwrap_or(usize::MAX);
-        let matched_id = cached.rule_ids.get(found.index).copied().unwrap_or(0);
-
-        let dominated = if let Some((_, _, best_start, best_prio)) = &best_result {
-            if start < *best_start {
-                false
-            } else if start == *best_start {
-                *best_prio == InjectionPriority::Left && *priority != InjectionPriority::Left
-            } else {
-                true
-            }
-        } else {
-            false
-        };
-
-        if !dominated {
-            best_result = Some((found, matched_id, start, *priority));
-            if start == cursor && *priority == InjectionPriority::Left {
-                break;
-            }
-        }
-    }
-
-    best_result.map(|(found, matched_id, _, prio)| (found, matched_id, prio))
-}
-
-// ─────────────────────────────────────────────────────────────────────────────
-// UTF-16 mapping and utility functions (unchanged)
-// ─────────────────────────────────────────────────────────────────────────────
-
-fn utf16_to_byte_map(input: &str) -> Vec<usize> {
-    let mut map = Vec::with_capacity(input.encode_utf16().count() + 1);
-    for (byte_idx, ch) in input.char_indices() {
-        map.push(byte_idx);
-        if ch.len_utf16() == 2 {
-            map.push(byte_idx);
-        }
-    }
-    map.push(input.len());
-    map
-}
-
-fn supports_plaintext(lang: &str) -> bool {
-    matches!(lang, "text" | "txt" | "plain" | "plaintext")
-}
-
-fn supports_json(lang: &str) -> bool {
-    lang == "json"
-}
-
-fn lang_mode_from_scope(scope_name: &str) -> Option<LangMode> {
-    if scope_name == "source.json" || scope_name.ends_with(".json") {
-        return Some(LangMode::Json);
-    }
-    if scope_name == "text.plain" {
-        return Some(LangMode::Plaintext);
-    }
-    None
-}
-
-fn resolve_lang_mode_from_lang(lang: &str) -> Option<LangMode> {
-    if supports_plaintext(lang) {
-        return Some(LangMode::Plaintext);
-    }
-    if supports_json(lang) {
-        return Some(LangMode::Json);
-    }
-    None
-}
-
-fn resolve_lang_from_options(options_json: &str) -> Result<String> {
-    let Some(lang) = parse_lang(options_json) else {
-        return Err(Error::from_reason(
-            "Ferriki vertical slice requires options.lang.",
-        ));
-    };
-    Ok(lang)
-}
-
-fn escape_html(input: &str) -> String {
-    let mut escaped = String::with_capacity(input.len());
-    for ch in input.chars() {
-        match ch {
-            '&' => escaped.push_str("&#x26;"),
-            '<' => escaped.push_str("&#x3C;"),
-            _ => escaped.push(ch),
-        }
-    }
-    escaped
-}
-
-fn render_plain_html(
-    code: &str,
-    options_json: &str,
-    themes: &HashMap<String, ThemeData>,
-) -> String {
-    let theme = resolve_html_theme_profile(options_json, "ferriki-plain", themes);
-    render_unstyled_html(code, &theme)
-}
-
-fn render_plain_tokens_json(
-    code: &str,
-    options_json: &str,
-    themes: &HashMap<String, ThemeData>,
-) -> Result<String> {
-    let theme = resolve_theme_profile(options_json, "ferriki-plain", themes);
-    let utf16_len = code.encode_utf16().count();
-    let styled = vec![StyledJsonToken {
-        content: code.to_owned(),
-        content_utf16_len: utf16_len,
-        offset_utf16: 0,
-        color: Arc::<str>::from(COLOR_DEFAULT_FG),
-        font_style: 0,
-        dark_color: None,
-    }];
-    let styled_lines = styled_json_lines(&styled);
-    let line_start_offsets = line_start_offsets_utf16(code);
-
-    let mut out = String::with_capacity(styled_lines.len() * 64);
-    out.push_str("{\"tokens\":[");
-    for (line_index, line) in styled_lines.into_iter().enumerate() {
-        if line_index > 0 {
-            out.push(',');
-        }
-        let line_offset = line_start_offsets
-            .get(line_index)
-            .copied()
-            .unwrap_or_else(|| line_start_offsets.last().copied().unwrap_or(0));
-
-        if line.is_empty() {
-            out.push_str("[{\"content\":\"\",\"offset\":");
-            push_usize(&mut out, line_offset);
-            out.push_str("}]");
-            continue;
-        }
-
-        let mut content = String::new();
-        let mut offset = line_offset;
-        for (index, token) in line.iter().enumerate() {
-            if index == 0 {
-                offset = token.offset_utf16;
-            }
-            content.push_str(&token.content);
-        }
-
-        out.push_str("[{\"content\":\"");
-        push_json_escaped(&mut out, &content);
-        out.push_str("\",\"offset\":");
-        push_usize(&mut out, offset);
-        out.push_str("}]");
-    }
-    out.push_str("],\"themeName\":\"");
-    push_json_escaped(&mut out, &theme.theme_name);
-    out.push_str("\",\"fg\":\"");
-    out.push_str(&theme.fg.unwrap_or_default());
-    out.push_str("\",\"bg\":\"");
-    out.push_str(&theme.bg.unwrap_or_default());
-    out.push_str("\"}");
-    Ok(out)
-}
-
-fn render_plain_hast_json(
-    code: &str,
-    options_json: &str,
-    themes: &HashMap<String, ThemeData>,
-) -> Result<String> {
-    let theme = resolve_html_theme_profile(options_json, "ferriki-plain", themes);
-    let utf16_len = code.encode_utf16().count();
-    let styled = vec![StyledJsonToken {
-        content: code.to_owned(),
-        content_utf16_len: utf16_len,
-        offset_utf16: 0,
-        color: Arc::<str>::from(COLOR_DEFAULT_FG),
-        font_style: 0,
-        dark_color: None,
-    }];
-    let lines = styled_json_lines(&styled);
-    render_styled_hast_payload_json(&lines, options_json, &theme, None)
-}
-
-fn line_start_offsets_utf16(input: &str) -> Vec<usize> {
-    let mut starts = vec![0usize];
-    let mut offset = 0usize;
-
-    for ch in input.chars() {
-        offset = offset.saturating_add(ch.len_utf16());
-        if ch == '\n' {
-            starts.push(offset);
-        }
-    }
-
-    starts
-}
-
-fn push_slice(
-    out: &mut Vec<JsonToken>,
-    kind: &'static str,
-    start_utf16: usize,
-    end_utf16: usize,
-    utf16_map: &[usize],
-    code: &str,
-) -> Result<()> {
-    if end_utf16 < start_utf16 || end_utf16 >= utf16_map.len() {
-        return Err(Error::from_reason(
-            "Ferriki JSON tokenizer produced invalid range.",
-        ));
-    }
-
-    let start_byte = utf16_map[start_utf16];
-    let end_byte = utf16_map[end_utf16];
-    let content = code
-        .get(start_byte..end_byte)
-        .ok_or_else(|| Error::from_reason("Ferriki JSON tokenizer failed to slice source text."))?
-        .to_owned();
-
-    out.push(JsonToken {
-        kind,
-        start_utf16,
-        end_utf16,
-        content,
-    });
-    Ok(())
-}
-
-fn tokenize_json_with_ferroni(code: &str) -> Result<Vec<JsonToken>> {
-    let patterns = [
-        r#""(?:\\.|[^"\\])*""#,
-        r"-?(?:0|[1-9][0-9]*)(?:\.[0-9]+)?(?:[eE][+-]?[0-9]+)?",
-        r"\b(?:true|false|null)\b",
-        r"[{}\[\]:,]",
-    ];
-    let pattern_refs = patterns.to_vec();
-    let mut scanner = Scanner::new(&pattern_refs).map_err(|err| {
-        Error::from_reason(format!("Failed to initialize Ferroni JSON scanner: {err}"))
-    })?;
-    let input = OnigString::new(code);
-    let utf16_map = utf16_to_byte_map(code);
-    let total_utf16 = utf16_map.len().saturating_sub(1);
-    let find_options = ScannerFindOptions::from_bits(0);
-
-    let mut cursor = 0usize;
-    let mut out = Vec::new();
-
-    while cursor < total_utf16 {
-        let Some(found) = scanner.find_next_match_utf16(&input, cursor, find_options) else {
-            push_slice(&mut out, "text", cursor, total_utf16, &utf16_map, code)?;
-            break;
-        };
-
-        let first = found.capture_indices.first().ok_or_else(|| {
-            Error::from_reason("Ferriki JSON scanner returned match without capture 0.")
-        })?;
-        let start_utf16 = first.start;
-        let end_utf16 = first.end;
-
-        if start_utf16 > cursor {
-            push_slice(&mut out, "text", cursor, start_utf16, &utf16_map, code)?;
-        }
-
-        if end_utf16 <= start_utf16 {
-            cursor = start_utf16.saturating_add(1);
-            continue;
-        }
-
-        let kind = match found.index {
-            0 => "string",
-            1 => "number",
-            2 => "literal",
-            3 => "punct",
-            _ => "text",
-        };
-
-        push_slice(&mut out, kind, start_utf16, end_utf16, &utf16_map, code)?;
-        cursor = end_utf16;
-    }
-
-    Ok(out)
-}
-
-fn merge_adjacent_json_punct_tokens(tokens: Vec<JsonToken>) -> Vec<JsonToken> {
-    let mut merged: Vec<JsonToken> = Vec::with_capacity(tokens.len());
-
-    for token in tokens {
-        if token.kind == "punct" {
-            if let Some(last) = merged.last_mut() {
-                if last.kind == "punct" && last.end_utf16 == token.start_utf16 {
-                    last.end_utf16 = token.end_utf16;
-                    last.content.push_str(&token.content);
-                    continue;
-                }
-            }
-        }
-
-        merged.push(token);
-    }
-
-    merged
-}
-
-fn theme_profile_by_name(
-    theme_name: &str,
-    themes: &HashMap<String, ThemeData>,
-) -> JsonThemeProfile {
-    if theme_name == "none" {
-        return JsonThemeProfile {
-            pre_class: "shiki none".to_owned(),
-            pre_style: Some("background-color:;color:".to_owned()),
-            theme_name: "none".to_owned(),
-            fg: None,
-            bg: None,
-        };
-    }
-
-    if let Some(theme_data) = themes.get(theme_name) {
-        let fg = if theme_data.fg.is_empty() {
-            None
-        } else {
-            Some(theme_data.fg.clone())
-        };
-        let bg = if theme_data.bg.is_empty() {
-            None
-        } else {
-            Some(theme_data.bg.clone())
-        };
-        let pre_style = match (&fg, &bg) {
-            (Some(f), Some(b)) => Some(format!("background-color:{b};color:{f}")),
-            _ => None,
-        };
-        return JsonThemeProfile {
-            pre_class: format!("shiki {theme_name}"),
-            pre_style,
-            theme_name: theme_name.to_owned(),
-            fg,
-            bg,
-        };
-    }
-
-    JsonThemeProfile {
-        pre_class: format!("shiki {theme_name}"),
-        pre_style: None,
-        theme_name: theme_name.to_owned(),
-        fg: None,
-        bg: None,
-    }
-}
-
-fn resolve_theme_profile(
-    options_json: &str,
-    fallback_theme: &str,
-    themes: &HashMap<String, ThemeData>,
-) -> JsonThemeProfile {
-    if let Some((light, _dark)) = parse_dual_themes(options_json) {
-        return theme_profile_by_name(&light, themes);
-    }
-    let theme_name = parse_theme(options_json).unwrap_or_else(|| fallback_theme.to_owned());
-    theme_profile_by_name(&theme_name, themes)
-}
-
-fn resolve_html_theme_profile(
-    options_json: &str,
-    fallback_theme: &str,
-    themes: &HashMap<String, ThemeData>,
-) -> HtmlThemeProfile {
-    if let Some((light, dark)) = parse_dual_themes(options_json) {
-        let light_profile = theme_profile_by_name(&light, themes);
-        let dark_profile = theme_profile_by_name(&dark, themes);
-        let light_bg = light_profile.bg.clone().unwrap_or_default();
-        let light_fg = light_profile.fg.clone().unwrap_or_default();
-        let dark_bg = if dark == "none" {
-            COLOR_INHERIT.to_owned()
-        } else {
-            dark_profile.bg.clone().unwrap_or_default()
-        };
-        let dark_fg = if dark == "none" {
-            COLOR_INHERIT.to_owned()
-        } else {
-            dark_profile.fg.clone().unwrap_or_default()
-        };
-
-        return HtmlThemeProfile {
-      pre_class: format!("shiki shiki-themes {light} {dark}"),
-      pre_style: Some(format!(
-        "background-color:{light_bg};--shiki-dark-bg:{dark_bg};color:{light_fg};--shiki-dark:{dark_fg}"
-      )),
-      theme_name: light_profile.theme_name,
-      dark_theme_name: Some(dark_profile.theme_name),
-      fg: Some(light_fg),
-      bg: Some(light_bg),
-      dark_fg: Some(dark_fg),
-      dark_bg: Some(dark_bg),
-      disable_token_coloring: light == "none",
-    };
-    }
-
-    let profile = resolve_theme_profile(options_json, fallback_theme, themes);
-    let disable_token_coloring = profile.theme_name == "none";
-    HtmlThemeProfile {
-        pre_class: profile.pre_class,
-        pre_style: profile.pre_style,
-        theme_name: profile.theme_name,
-        dark_theme_name: None,
-        fg: profile.fg,
-        bg: profile.bg,
-        dark_fg: None,
-        dark_bg: None,
-        disable_token_coloring,
-    }
-}
-
-fn resolve_json_theme(options_json: &str, themes: &HashMap<String, ThemeData>) -> JsonThemeProfile {
-    resolve_theme_profile(options_json, "ferriki-json", themes)
-}
-
-fn is_json_key_string(tokens: &[JsonToken], index: usize) -> bool {
-    for token in tokens.iter().skip(index.saturating_add(1)) {
-        if token.kind == "text" && token.content.chars().all(char::is_whitespace) {
-            continue;
-        }
-        return token.kind == "punct" && token.content == ":";
-    }
-    false
-}
-
-fn push_styled_token(
-    out: &mut Vec<StyledJsonToken>,
-    content: String,
-    offset_utf16: usize,
-    color: &Arc<str>,
-) {
-    if content.is_empty() {
-        return;
-    }
-    let content_utf16_len = content.encode_utf16().count();
-    out.push(StyledJsonToken {
-        content,
-        content_utf16_len,
-        offset_utf16,
-        color: color.clone(),
-        font_style: 0,
-        dark_color: None,
-    });
-}
-
-fn push_styled_string_token(
-    out: &mut Vec<StyledJsonToken>,
-    token: &JsonToken,
-    _is_key: bool,
-    quote_color: &Arc<str>,
-    body_color: &Arc<str>,
-) {
-    if quote_color == body_color {
-        push_styled_token(out, token.content.clone(), token.start_utf16, body_color);
-        return;
-    }
-
-    if token.content.len() >= 2 && token.content.starts_with('"') && token.content.ends_with('"') {
-        let char_count = token.content.chars().count();
-        if char_count >= 2 {
-            let body = token
-                .content
-                .chars()
-                .skip(1)
-                .take(char_count.saturating_sub(2))
-                .collect::<String>();
-            push_styled_token(out, "\"".to_owned(), token.start_utf16, quote_color);
-            push_styled_token(out, body, token.start_utf16.saturating_add(1), body_color);
-            push_styled_token(
-                out,
-                "\"".to_owned(),
-                token.end_utf16.saturating_sub(1),
-                quote_color,
-            );
-            return;
-        }
-    }
-
-    push_styled_token(out, token.content.clone(), token.start_utf16, body_color);
-}
-
-/// Write a JSON-escaped string (handles \n, \r, \t, \\, \", and control chars)
-fn push_json_escaped(out: &mut String, s: &str) {
-    for ch in s.chars() {
-        match ch {
-            '"' => out.push_str("\\\""),
-            '\\' => out.push_str("\\\\"),
-            '\n' => out.push_str("\\n"),
-            '\r' => out.push_str("\\r"),
-            '\t' => out.push_str("\\t"),
-            c if c < '\x20' => {
-                out.push_str(&format!("\\u{:04x}", c as u32));
-            }
-            c => out.push(c),
-        }
-    }
-}
-
-/// Write a usize as decimal digits without allocation
-fn push_usize(out: &mut String, n: usize) {
-    if n == 0 {
-        out.push('0');
-        return;
-    }
-    let start = out.len();
-    let mut val = n;
-    while val > 0 {
-        out.push((b'0' + (val % 10) as u8) as char);
-        val /= 10;
-    }
-    // Reverse the digits we just pushed
-    unsafe {
-        let bytes = out.as_bytes_mut();
-        bytes[start..].reverse();
-    }
-}
-
-fn normalize_hex_color(color: &str) -> String {
-    match color.strip_prefix('#') {
-        Some(rest) => format!("#{}", rest.to_uppercase()),
-        None => color.to_owned(),
-    }
-}
-
-fn resolve_json_scope_color(scope_names: &[&str], theme: &ThemeData) -> Arc<str> {
-    let style = resolve_token_style(scope_names, theme);
-    style
-        .foreground
-        .unwrap_or_else(|| theme.fg_normalized.clone())
-}
-
-fn resolve_json_scope_color_with_fallback(
-    scope_names: &[&str],
-    fallback_color: &Arc<str>,
-    theme: &ThemeData,
-) -> Arc<str> {
-    let mut has_specific_match = false;
-    for rule in &theme.settings {
-        if rule.scopes.is_empty() {
-            continue;
-        }
-        for parts in &rule.scope_parts {
-            if selector_matches_presplit(parts, scope_names).is_some() {
-                has_specific_match = true;
-                break;
-            }
-        }
-        if has_specific_match {
-            break;
-        }
-    }
-
-    if has_specific_match {
-        resolve_json_scope_color(scope_names, theme)
-    } else {
-        fallback_color.clone()
-    }
-}
-
-fn style_json_tokens(tokens: &[JsonToken], theme: &ThemeData) -> Vec<StyledJsonToken> {
-    let mut styled = Vec::new();
-    let root = "source.json";
-    let default_fg = theme.fg_normalized.clone();
-
-    // Pre-resolve all JSON scope colors once (avoids repeated theme lookups per token)
-    let key_body_color = resolve_json_scope_color(
-        &[root, "string.json", "support.type.property-name.json"],
-        theme,
-    );
-    let key_quote_color = resolve_json_scope_color_with_fallback(
-        &[
-            root,
-            "string.json",
-            "support.type.property-name.json",
-            "punctuation.support.type.property-name.json",
-        ],
-        &key_body_color,
-        theme,
-    );
-    let key_has_separate_quotes = key_quote_color != key_body_color;
-
-    let str_body_color = resolve_json_scope_color(&[root, "string.quoted.double.json"], theme);
-    let str_quote_color = resolve_json_scope_color_with_fallback(
-        &[
-            root,
-            "string.quoted.double.json",
-            "punctuation.definition.string.json",
-        ],
-        &str_body_color,
-        theme,
-    );
-    let str_has_separate_quotes = str_quote_color != str_body_color;
-
-    let number_color = resolve_json_scope_color(&[root, "constant.numeric.json"], theme);
-    let literal_color = resolve_json_scope_color(&[root, "constant.language.json"], theme);
-    let punct_sep_color = resolve_json_scope_color(&[root, "punctuation.separator.json"], theme);
-    let punct_def_color = resolve_json_scope_color(&[root, "punctuation.definition.json"], theme);
-
-    for (index, token) in tokens.iter().enumerate() {
-        match token.kind {
-            "text" => push_styled_token(
-                &mut styled,
-                token.content.clone(),
-                token.start_utf16,
-                &default_fg,
-            ),
-            "string" => {
-                let is_key = is_json_key_string(tokens, index);
-                if is_key {
-                    if key_has_separate_quotes {
-                        push_styled_string_token(
-                            &mut styled,
-                            token,
-                            true,
-                            &key_quote_color,
-                            &key_body_color,
-                        );
-                    } else {
-                        push_styled_token(
-                            &mut styled,
-                            token.content.clone(),
-                            token.start_utf16,
-                            &key_body_color,
-                        );
-                    }
-                } else {
-                    if str_has_separate_quotes {
-                        push_styled_string_token(
-                            &mut styled,
-                            token,
-                            false,
-                            &str_quote_color,
-                            &str_body_color,
-                        );
-                    } else {
-                        push_styled_token(
-                            &mut styled,
-                            token.content.clone(),
-                            token.start_utf16,
-                            &str_body_color,
-                        );
-                    }
-                }
-            }
-            "number" => {
-                push_styled_token(
-                    &mut styled,
-                    token.content.clone(),
-                    token.start_utf16,
-                    &number_color,
-                );
-            }
-            "literal" => {
-                push_styled_token(
-                    &mut styled,
-                    token.content.clone(),
-                    token.start_utf16,
-                    &literal_color,
-                );
-            }
-            "punct" => {
-                let is_sep = token.content == ":" || token.content == ",";
-                let color = if is_sep {
-                    &punct_sep_color
-                } else {
-                    &punct_def_color
-                };
-                push_styled_token(&mut styled, token.content.clone(), token.start_utf16, color);
-            }
-            _ => push_styled_token(
-                &mut styled,
-                token.content.clone(),
-                token.start_utf16,
-                &default_fg,
-            ),
-        }
-    }
-
-    styled
-}
-
-fn styled_json_lines(styled: &[StyledJsonToken]) -> Vec<Vec<StyledJsonToken>> {
-    let mut lines: Vec<Vec<StyledJsonToken>> = vec![Vec::new()];
-
-    for token in styled {
-        let mut piece = String::new();
-        let mut offset_utf16 = token.offset_utf16;
-        let mut piece_start_utf16 = token.offset_utf16;
-
-        for ch in token.content.chars() {
-            if ch == '\r' {
-                offset_utf16 = offset_utf16.saturating_add(1);
-                continue;
-            }
-
-            if ch == '\n' {
-                if !piece.is_empty() {
-                    lines
-                        .last_mut()
-                        .expect("line exists")
-                        .push(StyledJsonToken {
-                            content: piece.clone(),
-                            content_utf16_len: offset_utf16 - piece_start_utf16,
-                            offset_utf16: piece_start_utf16,
-                            color: token.color.clone(),
-                            font_style: token.font_style,
-                            dark_color: token.dark_color.clone(),
-                        });
-                }
-                piece.clear();
-                lines.push(Vec::new());
-                offset_utf16 = offset_utf16.saturating_add(1);
-                piece_start_utf16 = offset_utf16;
-                continue;
-            }
-
-            if piece.is_empty() {
-                piece_start_utf16 = offset_utf16;
-            }
-
-            piece.push(ch);
-            offset_utf16 = offset_utf16.saturating_add(ch.len_utf16());
-        }
-
-        if !piece.is_empty() {
-            lines
-                .last_mut()
-                .expect("line exists")
-                .push(StyledJsonToken {
-                    content: piece,
-                    content_utf16_len: offset_utf16 - piece_start_utf16,
-                    offset_utf16: piece_start_utf16,
-                    color: token.color.clone(),
-                    font_style: token.font_style,
-                    dark_color: token.dark_color.clone(),
-                });
-        }
-    }
-
-    lines
-}
-
-fn merge_line_for_html(line: &[StyledJsonToken], _default_fg: &str) -> Vec<StyledJsonToken> {
-    line.to_vec()
-}
-
-fn merge_leading_whitespace_tokens(line: &[StyledJsonToken]) -> Vec<StyledJsonToken> {
-    const FONT_STYLE_UNDERLINE: u8 = 4;
-    const FONT_STYLE_STRIKETHROUGH: u8 = 8;
-
-    let mut merged: Vec<StyledJsonToken> = Vec::with_capacity(line.len());
-    let mut carry: Option<StyledJsonToken> = None;
-
-    for token in line {
-        let content = token.content.as_str();
-        let is_decorated =
-            token.font_style & (FONT_STYLE_UNDERLINE | FONT_STYLE_STRIKETHROUGH) != 0;
-        let is_whitespace_only = !content.is_empty() && content.chars().all(char::is_whitespace);
-
-        if !is_decorated && is_whitespace_only {
-            if let Some(existing) = carry.as_mut() {
-                existing.content.push_str(content);
-                existing.content_utf16_len += token.content_utf16_len;
-            } else {
-                carry = Some(token.clone());
-            }
-            continue;
-        }
-
-        if let Some(carry_token) = carry.take() {
-            if !is_decorated {
-                let mut combined = token.clone();
-                combined.offset_utf16 = carry_token.offset_utf16;
-                combined.content = format!("{}{}", carry_token.content, token.content);
-                combined.content_utf16_len =
-                    carry_token.content_utf16_len + token.content_utf16_len;
-                merged.push(combined);
-            } else {
-                merged.push(carry_token);
-                merged.push(token.clone());
-            }
-            continue;
-        }
-
-        merged.push(token.clone());
-    }
-
-    if let Some(carry_token) = carry {
-        merged.push(carry_token);
-    }
-
-    merged
-}
-
-fn split_whitespace_tokens(line: &[StyledJsonToken]) -> Vec<StyledJsonToken> {
-    let mut split = Vec::with_capacity(line.len());
-
-    for token in line {
-        let content = token.content.as_str();
-        if !content.chars().any(char::is_whitespace) || content.chars().all(char::is_whitespace) {
-            split.push(token.clone());
-            continue;
-        }
-
-        let leading_len = content.chars().take_while(|ch| ch.is_whitespace()).count();
-        let trailing_len = content
-            .chars()
-            .rev()
-            .take_while(|ch| ch.is_whitespace())
-            .count();
-        if leading_len == 0 && trailing_len == 0 {
-            split.push(token.clone());
-            continue;
-        }
-
-        let total_len = content.chars().count();
-        let content_len = total_len.saturating_sub(leading_len + trailing_len);
-        let mut utf16_offset = token.offset_utf16;
-
-        if leading_len > 0 {
-            let leading: String = content.chars().take(leading_len).collect();
-            split.push(StyledJsonToken {
-                content_utf16_len: leading.encode_utf16().count(),
-                content: leading,
-                offset_utf16: utf16_offset,
-                color: Arc::<str>::from(""),
-                font_style: 0,
-                dark_color: None,
-            });
-            utf16_offset += split.last().expect("leading token").content_utf16_len;
-        }
-
-        if content_len > 0 {
-            let middle: String = content
-                .chars()
-                .skip(leading_len)
-                .take(content_len)
-                .collect();
-            split.push(StyledJsonToken {
-                content_utf16_len: middle.encode_utf16().count(),
-                content: middle,
-                offset_utf16: utf16_offset,
-                color: token.color.clone(),
-                font_style: token.font_style,
-                dark_color: token.dark_color.clone(),
-            });
-            utf16_offset += split.last().expect("middle token").content_utf16_len;
-        }
-
-        if trailing_len > 0 {
-            let trailing: String = content.chars().skip(leading_len + content_len).collect();
-            split.push(StyledJsonToken {
-                content_utf16_len: trailing.encode_utf16().count(),
-                content: trailing,
-                offset_utf16: utf16_offset,
-                color: Arc::<str>::from(""),
-                font_style: 0,
-                dark_color: None,
-            });
-        }
-    }
-
-    split
-}
-
-fn merge_adjacent_styled_tokens(line: &[StyledJsonToken]) -> Vec<StyledJsonToken> {
-    let mut merged: Vec<StyledJsonToken> = Vec::with_capacity(line.len());
-
-    for token in line {
-        if let Some(previous) = merged.last_mut() {
-            let prev_decorated = previous.font_style & 4 != 0 || previous.font_style & 8 != 0;
-            let current_decorated = token.font_style & 4 != 0 || token.font_style & 8 != 0;
-            if !prev_decorated
-                && !current_decorated
-                && previous.color == token.color
-                && previous.font_style == token.font_style
-                && previous.dark_color == token.dark_color
-            {
-                previous.content.push_str(&token.content);
-                previous.content_utf16_len += token.content_utf16_len;
-                continue;
-            }
-        }
-        merged.push(token.clone());
-    }
-
-    merged
-}
-
-fn parse_merge_whitespaces_mode(options: &Value) -> i8 {
-    match options.get("mergeWhitespaces") {
-        Some(Value::String(value)) if value == "never" => -1,
-        Some(Value::Bool(false)) => 0,
-        _ => 1,
-    }
-}
-
-fn parse_merge_same_style_tokens(options: &Value) -> bool {
-    matches!(options.get("mergeSameStyleTokens"), Some(Value::Bool(true)))
-}
-
-fn apply_render_line_options(
-    lines: Vec<Vec<StyledJsonToken>>,
-    options: &Value,
-) -> Vec<Vec<StyledJsonToken>> {
-    let merge_whitespace_mode = parse_merge_whitespaces_mode(options);
-    let merge_same_style = parse_merge_same_style_tokens(options);
-
-    lines
-        .into_iter()
-        .map(|line| {
-            let line = match merge_whitespace_mode {
-                -1 => split_whitespace_tokens(&line),
-                1 => merge_leading_whitespace_tokens(&line),
-                _ => line,
-            };
-            if merge_same_style {
-                merge_adjacent_styled_tokens(&line)
-            } else {
-                line
-            }
-        })
-        .collect()
-}
-
-fn collect_color_replacements_from_source(
-    target: &mut HashMap<String, String>,
-    source: Option<&Value>,
-    theme_name: &str,
-) {
-    let Some(Value::Object(map)) = source else {
-        return;
-    };
-
-    for (key, value) in map {
-        if let Some(replacement) = value.as_str() {
-            target.insert(normalize_hex_color(key), replacement.to_owned());
-            continue;
-        }
-
-        if key != theme_name {
-            continue;
-        }
-
-        let Some(scoped) = value.as_object() else {
-            continue;
-        };
-        for (scoped_key, scoped_value) in scoped {
-            if let Some(replacement) = scoped_value.as_str() {
-                target.insert(normalize_hex_color(scoped_key), replacement.to_owned());
-            }
-        }
-    }
-}
-
-fn resolve_color_replacements(options: &Value, theme_name: &str) -> HashMap<String, String> {
-    let mut replacements = HashMap::new();
-    collect_color_replacements_from_source(
-        &mut replacements,
-        options.get("colorReplacements"),
-        theme_name,
-    );
-    replacements
-}
-
-fn apply_color_replacement_value(color: &str, replacements: &HashMap<String, String>) -> String {
-    replacements
-        .get(&normalize_hex_color(color))
-        .cloned()
-        .unwrap_or_else(|| color.to_owned())
-}
-
-fn apply_color_replacements_to_lines(
-    lines: &mut [Vec<StyledJsonToken>],
-    replacements: &HashMap<String, String>,
-) {
-    if replacements.is_empty() {
-        return;
-    }
-
-    for line in lines {
-        for token in line {
-            token.color =
-                Arc::<str>::from(apply_color_replacement_value(&token.color, replacements));
-            if let Some(dark_color) = token.dark_color.as_ref() {
-                token.dark_color = Some(Arc::<str>::from(apply_color_replacement_value(
-                    dark_color,
-                    replacements,
-                )));
-            }
-        }
-    }
-}
-
-fn apply_color_replacements_to_json_theme_profile(
-    theme: &mut JsonThemeProfile,
-    replacements: &HashMap<String, String>,
-) {
-    if replacements.is_empty() {
-        return;
-    }
-
-    theme.fg = theme
-        .fg
-        .as_ref()
-        .map(|fg| apply_color_replacement_value(fg, replacements));
-    theme.bg = theme
-        .bg
-        .as_ref()
-        .map(|bg| apply_color_replacement_value(bg, replacements));
-    theme.pre_style = match (&theme.fg, &theme.bg) {
-        (Some(fg), Some(bg)) => Some(format!("background-color:{bg};color:{fg}")),
-        _ => theme.pre_style.clone(),
-    };
-}
-
-fn apply_color_replacements_to_html_theme_profile(
-    theme: &mut HtmlThemeProfile,
-    replacements: &HashMap<String, String>,
-) {
-    if replacements.is_empty() {
-        return;
-    }
-
-    theme.fg = theme
-        .fg
-        .as_ref()
-        .map(|fg| apply_color_replacement_value(fg, replacements));
-    theme.bg = theme
-        .bg
-        .as_ref()
-        .map(|bg| apply_color_replacement_value(bg, replacements));
-    theme.dark_fg = theme
-        .dark_fg
-        .as_ref()
-        .map(|fg| apply_color_replacement_value(fg, replacements));
-    theme.dark_bg = theme
-        .dark_bg
-        .as_ref()
-        .map(|bg| apply_color_replacement_value(bg, replacements));
-
-    theme.pre_style = if let (Some(fg), Some(bg)) = (&theme.fg, &theme.bg) {
-        if let (Some(dark_fg), Some(dark_bg)) = (&theme.dark_fg, &theme.dark_bg) {
-            Some(format!(
-                "background-color:{bg};--shiki-dark-bg:{dark_bg};color:{fg};--shiki-dark:{dark_fg}"
-            ))
-        } else {
-            Some(format!("background-color:{bg};color:{fg}"))
-        }
-    } else {
-        theme.pre_style.clone()
-    };
-}
-
-fn apply_dark_theme_inherit(mut styled: Vec<StyledJsonToken>) -> Vec<StyledJsonToken> {
-    for token in &mut styled {
-        token.dark_color = Some(Arc::<str>::from(COLOR_INHERIT));
-    }
-    styled
-}
-
-fn apply_dark_theme_palette(
-    mut light_styled: Vec<StyledJsonToken>,
-    dark_styled: &[StyledJsonToken],
-) -> Vec<StyledJsonToken> {
-    for (index, light) in light_styled.iter_mut().enumerate() {
-        let Some(dark) = dark_styled.get(index) else {
-            break;
-        };
-        if dark.offset_utf16 == light.offset_utf16 && dark.content == light.content {
-            light.dark_color = Some(dark.color.clone());
-        }
-    }
-    light_styled
-}
-
-fn render_unstyled_html(code: &str, theme: &HtmlThemeProfile) -> String {
-    let utf16_len = code.encode_utf16().count();
-    let styled = vec![StyledJsonToken {
-        content: code.to_owned(),
-        content_utf16_len: utf16_len,
-        offset_utf16: 0,
-        color: Arc::<str>::from(COLOR_DEFAULT_FG),
-        font_style: 0,
-        dark_color: None,
-    }];
-    let lines = styled_json_lines(&styled);
-    render_styled_html_lines(&lines, theme, true)
-}
-
-fn render_styled_html_lines(
-    lines: &[Vec<StyledJsonToken>],
-    theme: &HtmlThemeProfile,
-    unstyled_spans: bool,
-) -> String {
-    let mut html = String::new();
-    html.push_str("<pre class=\"");
-    html.push_str(&theme.pre_class);
-    html.push('"');
-    if let Some(style) = &theme.pre_style {
-        html.push_str(" style=\"");
-        html.push_str(style);
-        html.push('"');
-    }
-    html.push_str(" tabindex=\"0\"><code>");
-
-    for (line_index, line) in lines.iter().enumerate() {
-        html.push_str("<span class=\"line\">");
-        if line.is_empty() {
-            if unstyled_spans {
-                html.push_str("<span></span>");
-            }
-        } else {
-            for token in line {
-                if unstyled_spans {
-                    html.push_str("<span>");
-                } else if let Some(style) = styled_token_style_string(token) {
-                    html.push_str("<span style=\"");
-                    html.push_str(&style);
-                    html.push_str("\">");
-                } else {
-                    html.push_str("<span>");
-                }
-                html.push_str(&escape_html(&token.content));
-                html.push_str("</span>");
-            }
-        }
-        html.push_str("</span>");
-        if line_index + 1 < lines.len() {
-            html.push('\n');
-        }
-    }
-
-    html.push_str("</code></pre>");
-    html
-}
-
-fn styled_token_style_string(token: &StyledJsonToken) -> Option<String> {
-    let mut style = String::new();
-
-    if !token.color.is_empty() {
-        style.push_str("color:");
-        style.push_str(&token.color);
-    }
-    if let Some(dark_color) = &token.dark_color {
-        if !style.is_empty() {
-            style.push(';');
-        }
-        style.push_str("--shiki-dark:");
-        style.push_str(dark_color);
-    }
-    if token.font_style & 1 != 0 {
-        if !style.is_empty() {
-            style.push(';');
-        }
-        style.push_str("font-style:italic");
-    }
-    if token.font_style & 2 != 0 {
-        if !style.is_empty() {
-            style.push(';');
-        }
-        style.push_str("font-weight:bold");
-    }
-    if token.font_style & 4 != 0 || token.font_style & 8 != 0 {
-        if !style.is_empty() {
-            style.push(';');
-        }
-        style.push_str("text-decoration:");
-        if token.font_style & 4 != 0 {
-            style.push_str("underline");
-        }
-        if token.font_style & 8 != 0 {
-            if token.font_style & 4 != 0 {
-                style.push(' ');
-            }
-            style.push_str("line-through");
-        }
-    }
-
-    if style.is_empty() {
-        None
-    } else {
-        Some(style)
-    }
-}
-
-fn hast_text_node(value: &str) -> Value {
-    json!({
-      "type": "text",
-      "value": value,
-    })
-}
-
-fn hast_element_node(
-    tag_name: &str,
-    properties: serde_json::Map<String, Value>,
-    children: Vec<Value>,
-    data: Option<Value>,
-) -> Value {
-    let mut node = serde_json::Map::new();
-    node.insert("type".to_owned(), Value::String("element".to_owned()));
-    node.insert("tagName".to_owned(), Value::String(tag_name.to_owned()));
-    node.insert("properties".to_owned(), Value::Object(properties));
-    node.insert("children".to_owned(), Value::Array(children));
-    if let Some(data_value) = data {
-        node.insert("data".to_owned(), data_value);
-    }
-    Value::Object(node)
-}
-
-fn parse_hast_structure(options: &Value) -> &'static str {
-    match options.get("structure").and_then(Value::as_str) {
-        Some("inline") => "inline",
-        _ => "classic",
-    }
-}
-
-fn parse_hast_tabindex(options: &Value) -> Option<String> {
-    match options.get("tabindex") {
-        Some(Value::Bool(false)) | Some(Value::Null) => None,
-        Some(Value::String(value)) => Some(value.clone()),
-        Some(Value::Number(value)) => Some(value.to_string()),
-        Some(Value::Bool(value)) => Some(value.to_string()),
-        _ => Some("0".to_owned()),
-    }
-}
-
-fn resolve_hast_root_style(options: &Value, theme: &HtmlThemeProfile) -> Option<String> {
-    match options.get("rootStyle") {
-        Some(Value::Bool(false)) => None,
-        Some(Value::String(value)) => Some(value.clone()),
-        Some(Value::Null) => theme.pre_style.clone(),
-        _ => theme.pre_style.clone(),
-    }
-}
-
-fn render_styled_hast_payload_json(
-    lines: &[Vec<StyledJsonToken>],
-    options_json: &str,
-    theme: &HtmlThemeProfile,
-    rust_state: Option<Value>,
-) -> Result<String> {
-    let options: Value = serde_json::from_str(options_json).map_err(|err| {
-        Error::from_reason(format!("Failed to parse codeToHast options JSON: {err}"))
-    })?;
-    let structure = parse_hast_structure(&options);
-    let tabindex = parse_hast_tabindex(&options);
-    let root_style = resolve_hast_root_style(&options, theme);
-
-    let mut pre_properties = serde_json::Map::new();
-    pre_properties.insert("class".to_owned(), Value::String(theme.pre_class.clone()));
-    if let Some(style) = root_style {
-        pre_properties.insert("style".to_owned(), Value::String(style));
-    }
-    if let Some(tabindex) = tabindex {
-        pre_properties.insert("tabindex".to_owned(), Value::String(tabindex));
-    }
-    if let Some(meta) = options.get("meta").and_then(Value::as_object) {
-        for (key, value) in meta {
-            if !key.starts_with('_') {
-                pre_properties.insert(key.clone(), value.clone());
-            }
-        }
-    }
-
-    let mut root_children = Vec::new();
-    let mut code_children = Vec::new();
-
-    for (line_index, line) in lines.iter().enumerate() {
-        if line_index > 0 {
-            if structure == "inline" {
-                root_children.push(hast_element_node(
-                    "br",
-                    serde_json::Map::new(),
-                    Vec::new(),
-                    None,
-                ));
-            } else {
-                code_children.push(hast_text_node("\n"));
-            }
-        }
-
-        let mut line_children = Vec::new();
-        for token in line {
-            let mut token_properties = serde_json::Map::new();
-            if let Some(style) = styled_token_style_string(token) {
-                token_properties.insert("style".to_owned(), Value::String(style));
-            }
-            let token_node = hast_element_node(
-                "span",
-                token_properties,
-                vec![hast_text_node(&token.content)],
-                None,
-            );
-            if structure == "inline" {
-                root_children.push(token_node);
-            } else {
-                line_children.push(token_node);
-            }
-        }
-
-        if structure == "classic" {
-            let mut line_properties = serde_json::Map::new();
-            line_properties.insert("class".to_owned(), Value::String("line".to_owned()));
-            code_children.push(hast_element_node(
-                "span",
-                line_properties,
-                line_children,
-                None,
-            ));
-        }
-    }
-
-    if structure == "classic" {
-        let code_node = hast_element_node("code", serde_json::Map::new(), code_children, None);
-        let pre_node = hast_element_node(
-            "pre",
-            pre_properties,
-            vec![code_node],
-            options.get("data").cloned(),
-        );
-        root_children.push(pre_node);
-    }
-
-    let mut root_node = serde_json::Map::new();
-    root_node.insert("type".to_owned(), Value::String("root".to_owned()));
-    root_node.insert("children".to_owned(), Value::Array(root_children));
-
-    let mut payload = serde_json::Map::new();
-    payload.insert("hast".to_owned(), Value::Object(root_node));
-    if let Some(rust_state) = rust_state {
-        payload.insert("_rustState".to_owned(), rust_state);
-    }
-
-    serde_json::to_string(&Value::Object(payload))
-        .map_err(|err| Error::from_reason(format!("Failed to serialize codeToHast payload: {err}")))
-}
-
-fn render_styled_tokens_json(
-    lines: Vec<Vec<StyledJsonToken>>,
-    theme: JsonThemeProfile,
-) -> Result<String> {
-    // Manual JSON construction — avoids serde_json::Value heap allocations
-    let mut out = String::with_capacity(lines.len() * 128);
-    out.push_str("{\"tokens\":[");
-    for (li, line) in lines.iter().enumerate() {
-        if li > 0 {
-            out.push(',');
-        }
-        out.push('[');
-        for (ti, token) in line.iter().enumerate() {
-            if ti > 0 {
-                out.push(',');
-            }
-            out.push_str("{\"content\":\"");
-            push_json_escaped(&mut out, &token.content);
-            out.push_str("\",\"offset\":");
-            push_usize(&mut out, token.offset_utf16);
-            out.push_str(",\"color\":\"");
-            out.push_str(&token.color); // pre-normalized hex, no escaping needed
-            out.push_str("\",\"fontStyle\":");
-            push_usize(&mut out, token.font_style as usize);
-            out.push('}');
-        }
-        out.push(']');
-    }
-    out.push_str("],\"themeName\":\"");
-    push_json_escaped(&mut out, &theme.theme_name);
-    out.push('"');
-    if let Some(fg) = &theme.fg {
-        out.push_str(",\"fg\":\"");
-        out.push_str(fg);
-        out.push('"');
-    }
-    if let Some(bg) = &theme.bg {
-        out.push_str(",\"bg\":\"");
-        out.push_str(bg);
-        out.push('"');
-    }
-    out.push('}');
-    Ok(out)
-}
-
-fn render_styled_tokens_json_with_state(
-    lines: Vec<Vec<StyledJsonToken>>,
-    theme: JsonThemeProfile,
-    final_stack: &[StateFrame],
-    root_scope: Option<&str>,
-) -> Result<String> {
-    let mut out = String::with_capacity(lines.len() * 128);
-    out.push_str("{\"tokens\":[");
-    for (li, line) in lines.iter().enumerate() {
-        if li > 0 {
-            out.push(',');
-        }
-        out.push('[');
-        for (ti, token) in line.iter().enumerate() {
-            if ti > 0 {
-                out.push(',');
-            }
-            out.push_str("{\"content\":\"");
-            push_json_escaped(&mut out, &token.content);
-            out.push_str("\",\"offset\":");
-            push_usize(&mut out, token.offset_utf16);
-            out.push_str(",\"color\":\"");
-            out.push_str(&token.color);
-            out.push_str("\",\"fontStyle\":");
-            push_usize(&mut out, token.font_style as usize);
-            out.push('}');
-        }
-        out.push(']');
-    }
-    out.push_str("],\"themeName\":\"");
-    push_json_escaped(&mut out, &theme.theme_name);
-    out.push('"');
-    if let Some(fg) = &theme.fg {
-        out.push_str(",\"fg\":\"");
-        out.push_str(fg);
-        out.push('"');
-    }
-    if let Some(bg) = &theme.bg {
-        out.push_str(",\"bg\":\"");
-        out.push_str(bg);
-        out.push('"');
-    }
-    // Serialize state via serde (complex nested structure)
-    let state_value = serialize_state_frames(final_stack, root_scope);
-    out.push_str(",\"_rustState\":");
-    let state_json = serde_json::to_string(&state_value)
-        .map_err(|err| Error::from_reason(format!("Failed to serialize state: {err}")))?;
-    out.push_str(&state_json);
-    out.push('}');
-    Ok(out)
-}
-
-fn default_theme_data(theme_name: &str) -> ThemeData {
-    let fg = COLOR_DEFAULT_FG.to_owned();
-    ThemeData {
-        name: theme_name.to_owned(),
-        fg_normalized: Arc::<str>::from(COLOR_DEFAULT_FG),
-        fg,
-        bg: COLOR_DEFAULT_BG.to_owned(),
-        settings: Vec::new(),
-    }
-}
-
-fn resolve_theme_data<'a>(
-    theme_name: &str,
-    themes: &'a HashMap<String, ThemeData>,
-) -> Option<&'a ThemeData> {
-    themes.get(theme_name)
-}
-
-fn normalize_vue_tag_end_tokens(
-    lines: &mut [Vec<StyledJsonToken>],
-    root_scope: Option<&str>,
-    theme: &ThemeData,
-) {
-    if root_scope != Some("text.html.vue") {
-        return;
-    }
-
-    let punctuation = resolve_token_style(
-        &["text.html.vue", "punctuation.definition.tag.end.html.vue"],
-        theme,
-    )
-    .foreground
-    .unwrap_or_else(|| Arc::<str>::from("#666666"));
-
-    for line in lines.iter_mut() {
-        for token in line.iter_mut() {
-            if (token.content == ">" || token.content == "/>") && token.color == theme.fg_normalized
-            {
-                token.color = punctuation.clone();
-            }
-        }
-    }
-}
-
-fn push_normalized_token(
-    out: &mut Vec<StyledJsonToken>,
-    content: &str,
-    offset_utf16: usize,
-    color: Arc<str>,
-    font_style: u8,
-) {
-    if content.is_empty() {
-        return;
-    }
-    out.push(StyledJsonToken {
-        content: content.to_owned(),
-        content_utf16_len: content.encode_utf16().count(),
-        offset_utf16,
-        color,
-        font_style,
-        dark_color: None,
-    });
-}
-
-fn retokenize_astro_default_token(
-    token: &StyledJsonToken,
-    theme: &ThemeData,
-) -> Option<Vec<StyledJsonToken>> {
-    if token.color != theme.fg_normalized
-        || (!token.content.contains('<') && !token.content.contains('{'))
-    {
-        return None;
-    }
-
-    let punctuation = resolve_token_style(
-        &["source.astro", "punctuation.definition.tag.begin.astro"],
-        theme,
-    )
-    .foreground
-    .unwrap_or_else(|| Arc::<str>::from("#666666"));
-    let tag_name = resolve_token_style(&["source.astro", "entity.name.tag.astro"], theme)
-        .foreground
-        .unwrap_or_else(|| Arc::<str>::from("#4D9375"));
-    let identifier = resolve_token_style(&["source.astro", "identifier"], theme)
-        .foreground
-        .unwrap_or_else(|| Arc::<str>::from("#BD976A"));
-
-    let bytes = token.content.as_bytes();
-    let mut pieces = Vec::new();
-    let mut index = 0usize;
-    let mut offset = token.offset_utf16;
-    let mut in_interpolation = false;
-
-    while index < bytes.len() {
-        if bytes[index] == b'<' {
-            if index + 1 < bytes.len() && bytes[index + 1] == b'/' {
-                push_normalized_token(
-                    &mut pieces,
-                    "</",
-                    offset,
-                    punctuation.clone(),
-                    token.font_style,
-                );
-                index += 2;
-                offset += 2;
-            } else {
-                push_normalized_token(
-                    &mut pieces,
-                    "<",
-                    offset,
-                    punctuation.clone(),
-                    token.font_style,
-                );
-                index += 1;
-                offset += 1;
-            }
-
-            let start = index;
-            while index < bytes.len() {
-                let ch = bytes[index] as char;
-                if ch.is_ascii_alphanumeric() || matches!(ch, '-' | '_' | ':') {
-                    index += 1;
-                } else {
-                    break;
-                }
-            }
-            if index > start {
-                let name = &token.content[start..index];
-                push_normalized_token(
-                    &mut pieces,
-                    name,
-                    offset,
-                    tag_name.clone(),
-                    token.font_style,
-                );
-                offset += name.encode_utf16().count();
-            }
-            continue;
-        }
-
-        if bytes[index] == b'>' {
-            push_normalized_token(
-                &mut pieces,
-                ">",
-                offset,
-                punctuation.clone(),
-                token.font_style,
-            );
-            index += 1;
-            offset += 1;
-            continue;
-        }
-
-        if bytes[index] == b'{' {
-            in_interpolation = true;
-            push_normalized_token(
-                &mut pieces,
-                "{",
-                offset,
-                punctuation.clone(),
-                token.font_style,
-            );
-            index += 1;
-            offset += 1;
-            continue;
-        }
-
-        if bytes[index] == b'}' {
-            in_interpolation = false;
-            push_normalized_token(
-                &mut pieces,
-                "}",
-                offset,
-                punctuation.clone(),
-                token.font_style,
-            );
-            index += 1;
-            offset += 1;
-            continue;
-        }
-
-        if in_interpolation && (bytes[index] as char).is_ascii_alphabetic() {
-            let start = index;
-            while index < bytes.len() {
-                let ch = bytes[index] as char;
-                if ch.is_ascii_alphanumeric() || matches!(ch, '_' | '$') {
-                    index += 1;
-                } else {
-                    break;
-                }
-            }
-            let ident = &token.content[start..index];
-            push_normalized_token(
-                &mut pieces,
-                ident,
-                offset,
-                identifier.clone(),
-                token.font_style,
-            );
-            offset += ident.encode_utf16().count();
-            continue;
-        }
-
-        let start = index;
-        while index < bytes.len() && !matches!(bytes[index], b'<' | b'>' | b'{' | b'}') {
-            if in_interpolation && (bytes[index] as char).is_ascii_alphabetic() {
-                break;
-            }
-            index += 1;
-        }
-        let text = &token.content[start..index];
-        push_normalized_token(
-            &mut pieces,
-            text,
-            offset,
-            token.color.clone(),
-            token.font_style,
-        );
-        offset += text.encode_utf16().count();
-    }
-
-    Some(pieces)
-}
-
-fn recolor_astro_contextual_tokens(
-    line: &[StyledJsonToken],
-    theme: &ThemeData,
-) -> Vec<StyledJsonToken> {
-    let punctuation = resolve_token_style(
-        &["source.astro", "punctuation.definition.tag.begin.astro"],
-        theme,
-    )
-    .foreground
-    .unwrap_or_else(|| Arc::<str>::from("#666666"));
-    let tag_name = resolve_token_style(&["source.astro", "entity.name.tag.astro"], theme)
-        .foreground
-        .unwrap_or_else(|| Arc::<str>::from("#4D9375"));
-    let identifier = resolve_token_style(&["source.astro", "identifier"], theme)
-        .foreground
-        .unwrap_or_else(|| Arc::<str>::from("#BD976A"));
-
-    let mut normalized = Vec::with_capacity(line.len() + 2);
-    let mut expect_tag_name = false;
-    let mut expect_identifier = false;
-
-    for token in line {
-        let mut handled = false;
-        let mut remaining = token.content.as_str();
-        let mut offset = token.offset_utf16;
-
-        while !remaining.is_empty() {
-            if token.color == theme.fg_normalized && expect_identifier {
-                let ident_len = remaining
-                    .chars()
-                    .take_while(|ch| ch.is_ascii_alphanumeric() || matches!(ch, '_' | '$'))
-                    .map(char::len_utf8)
-                    .sum::<usize>();
-                if ident_len > 0 {
-                    let ident = &remaining[..ident_len];
-                    push_normalized_token(
-                        &mut normalized,
-                        ident,
-                        offset,
-                        identifier.clone(),
-                        token.font_style,
-                    );
-                    offset += ident.encode_utf16().count();
-                    remaining = &remaining[ident_len..];
-                    expect_identifier = false;
-                    handled = true;
-                    continue;
-                }
-                expect_identifier = false;
-            }
-
-            if token.color == theme.fg_normalized && expect_tag_name {
-                let tag_len = remaining
-                    .chars()
-                    .take_while(|ch| ch.is_ascii_alphanumeric() || matches!(ch, '-' | '_' | ':'))
-                    .map(char::len_utf8)
-                    .sum::<usize>();
-                if tag_len > 0 {
-                    let tag = &remaining[..tag_len];
-                    push_normalized_token(
-                        &mut normalized,
-                        tag,
-                        offset,
-                        tag_name.clone(),
-                        token.font_style,
-                    );
-                    offset += tag.encode_utf16().count();
-                    remaining = &remaining[tag_len..];
-                    expect_tag_name = false;
-                    handled = true;
-                    continue;
-                }
-                expect_tag_name = false;
-            }
-
-            if handled {
-                push_normalized_token(
-                    &mut normalized,
-                    remaining,
-                    offset,
-                    token.color.clone(),
-                    token.font_style,
-                );
-            } else {
-                normalized.push(token.clone());
-            }
-            break;
-        }
-
-        let current = normalized.last().unwrap_or(token);
-        let ends_with_punctuation = current.color == punctuation
-            && (current.content.ends_with('{')
-                || current.content.ends_with('<')
-                || current.content.ends_with("</"));
-        if ends_with_punctuation {
-            expect_identifier = current.content.ends_with('{');
-            expect_tag_name = current.content.ends_with('<') || current.content.ends_with("</");
-        }
-    }
-
-    normalized
-}
-
-fn merge_astro_punctuation_sequences(
-    line: &[StyledJsonToken],
-    theme: &ThemeData,
-) -> Vec<StyledJsonToken> {
-    let punctuation = resolve_token_style(
-        &["source.astro", "punctuation.definition.tag.begin.astro"],
-        theme,
-    )
-    .foreground
-    .unwrap_or_else(|| Arc::<str>::from("#666666"));
-
-    let mut merged: Vec<StyledJsonToken> = Vec::with_capacity(line.len());
-
-    for token in line {
-        if let Some(previous) = merged.last_mut() {
-            let is_contiguous =
-                previous.offset_utf16 + previous.content_utf16_len == token.offset_utf16;
-            let same_style = previous.color == punctuation
-                && token.color == punctuation
-                && previous.font_style == token.font_style
-                && previous.dark_color == token.dark_color;
-            let mergeable_sequence = matches!(
-                (previous.content.as_str(), token.content.as_str()),
-                (">", "{") | ("}", "</")
-            );
-            if is_contiguous && same_style && mergeable_sequence {
-                previous.content.push_str(&token.content);
-                previous.content_utf16_len += token.content_utf16_len;
-                continue;
-            }
-        }
-        merged.push(token.clone());
-    }
-
-    merged
-}
-
-fn normalize_astro_tag_tokens(
-    lines: &mut [Vec<StyledJsonToken>],
-    root_scope: Option<&str>,
-    theme: &ThemeData,
-) {
-    if root_scope != Some("source.astro") {
-        return;
-    }
-
-    for line in lines.iter_mut() {
-        let mut normalized = Vec::with_capacity(line.len() + 2);
-        for token in line.iter() {
-            if let Some(parts) = retokenize_astro_default_token(token, theme) {
-                normalized.extend(parts);
-            } else {
-                normalized.push(token.clone());
-            }
-        }
-        let recolored = recolor_astro_contextual_tokens(&normalized, theme);
-        *line = merge_astro_punctuation_sequences(&recolored, theme);
-    }
-}
-
-fn render_json_html(
-    code: &str,
-    options_json: &str,
-    themes: &HashMap<String, ThemeData>,
-) -> Result<String> {
-    let html_theme = resolve_html_theme_profile(options_json, "ferriki-json", themes);
-    if html_theme.disable_token_coloring {
-        return Ok(render_unstyled_html(code, &html_theme));
-    }
-
-    let fallback_light = default_theme_data(&html_theme.theme_name);
-    let light_theme = resolve_theme_data(&html_theme.theme_name, themes).unwrap_or(&fallback_light);
-    let options: Value = serde_json::from_str(options_json).map_err(|err| {
-        Error::from_reason(format!("Failed to parse codeToHtml options JSON: {err}"))
-    })?;
-    let replacements = resolve_color_replacements(&options, &html_theme.theme_name);
-    let mut html_theme = html_theme;
-    apply_color_replacements_to_html_theme_profile(&mut html_theme, &replacements);
-    let tokens = merge_adjacent_json_punct_tokens(tokenize_json_with_ferroni(code)?);
-    let mut styled = style_json_tokens(&tokens, light_theme);
-    if let Some(dark_theme_name) = html_theme.dark_theme_name.as_deref() {
-        if dark_theme_name == "none" {
-            styled = apply_dark_theme_inherit(styled);
-        } else {
-            let fallback_dark = default_theme_data(dark_theme_name);
-            let dark_theme = resolve_theme_data(dark_theme_name, themes).unwrap_or(&fallback_dark);
-            let dark_styled = style_json_tokens(&tokens, dark_theme);
-            styled = apply_dark_theme_palette(styled, &dark_styled);
-        }
-    }
-    let default_fg = light_theme.fg.clone();
-    let mut lines = styled_json_lines(&styled);
-    apply_color_replacements_to_lines(&mut lines, &replacements);
-    let lines = apply_render_line_options(lines, &options)
-        .into_iter()
-        .map(|line| merge_line_for_html(&line, &default_fg))
-        .collect::<Vec<_>>();
-    Ok(render_styled_html_lines(&lines, &html_theme, false))
-}
-
-fn render_json_tokens_json(
-    code: &str,
-    options_json: &str,
-    themes: &HashMap<String, ThemeData>,
-) -> Result<String> {
-    let options: Value = serde_json::from_str(options_json).map_err(|err| {
-        Error::from_reason(format!("Failed to parse codeToTokens options JSON: {err}"))
-    })?;
-    let tokens = merge_adjacent_json_punct_tokens(tokenize_json_with_ferroni(code)?);
-    let mut theme = resolve_json_theme(options_json, themes);
-    let fallback = default_theme_data(&theme.theme_name);
-    let theme_data = resolve_theme_data(&theme.theme_name, themes).unwrap_or(&fallback);
-    let replacements = resolve_color_replacements(&options, &theme.theme_name);
-    apply_color_replacements_to_json_theme_profile(&mut theme, &replacements);
-    let styled = style_json_tokens(&tokens, theme_data);
-    let mut lines = styled_json_lines(&styled);
-    apply_color_replacements_to_lines(&mut lines, &replacements);
-    render_styled_tokens_json(lines, theme)
-}
-
-fn render_json_hast_json(
-    code: &str,
-    options_json: &str,
-    themes: &HashMap<String, ThemeData>,
-) -> Result<String> {
-    let html_theme = resolve_html_theme_profile(options_json, "ferriki-json", themes);
-    let fallback_light = default_theme_data(&html_theme.theme_name);
-    let light_theme = resolve_theme_data(&html_theme.theme_name, themes).unwrap_or(&fallback_light);
-    let options: Value = serde_json::from_str(options_json).map_err(|err| {
-        Error::from_reason(format!("Failed to parse codeToHast options JSON: {err}"))
-    })?;
-    let replacements = resolve_color_replacements(&options, &html_theme.theme_name);
-    let mut html_theme = html_theme;
-    apply_color_replacements_to_html_theme_profile(&mut html_theme, &replacements);
-    let tokens = merge_adjacent_json_punct_tokens(tokenize_json_with_ferroni(code)?);
-    let mut styled = style_json_tokens(&tokens, light_theme);
-    if let Some(dark_theme_name) = html_theme.dark_theme_name.as_deref() {
-        if dark_theme_name == "none" {
-            styled = apply_dark_theme_inherit(styled);
-        } else {
-            let fallback_dark = default_theme_data(dark_theme_name);
-            let dark_theme = resolve_theme_data(dark_theme_name, themes).unwrap_or(&fallback_dark);
-            let dark_styled = style_json_tokens(&tokens, dark_theme);
-            styled = apply_dark_theme_palette(styled, &dark_styled);
-        }
-    }
-    let mut lines = styled_json_lines(&styled);
-    apply_color_replacements_to_lines(&mut lines, &replacements);
-    let lines = apply_render_line_options(lines, &options);
-    render_styled_hast_payload_json(&lines, options_json, &html_theme, None)
-}
-
-fn resolve_initial_stack(
-    options_json: &str,
-    code: &str,
-    compiled: &mut CompiledGrammar,
-    root_scope: Option<&str>,
-    theme: &ThemeData,
-) -> Result<Option<Vec<StateFrame>>> {
-    // Priority: _rustState > grammarContextCode > default (None)
-    if let Some(stack) = parse_initial_state_from_options(options_json) {
-        return Ok(Some(stack));
-    }
-    if let Some(context_code) = parse_grammar_context_code(options_json) {
-        if !context_code.is_empty() {
-            let (_, final_stack) =
-                tokenize_with_grammar_skeleton(&context_code, compiled, root_scope, theme, None)?;
-            return Ok(Some(final_stack));
-        }
-    }
-    let _ = code; // suppress unused warning
-    Ok(None)
-}
-
-fn render_grammar_html(
-    code: &str,
-    options_json: &str,
-    compiled: &mut CompiledGrammar,
-    root_scope: Option<&str>,
-    themes: &HashMap<String, ThemeData>,
-) -> Result<String> {
-    let options: Value = serde_json::from_str(options_json).map_err(|err| {
-        Error::from_reason(format!("Failed to parse codeToHtml options JSON: {err}"))
-    })?;
-    let html_theme = resolve_html_theme_profile(options_json, "ferriki-grammar", themes);
-    if html_theme.disable_token_coloring {
-        return Ok(render_unstyled_html(code, &html_theme));
-    }
-
-    let fallback_light = default_theme_data(&html_theme.theme_name);
-    let light_theme = resolve_theme_data(&html_theme.theme_name, themes).unwrap_or(&fallback_light);
-    let replacements = resolve_color_replacements(&options, &html_theme.theme_name);
-    let mut html_theme = html_theme;
-    apply_color_replacements_to_html_theme_profile(&mut html_theme, &replacements);
-    let initial_stack =
-        resolve_initial_stack(options_json, code, compiled, root_scope, light_theme)?;
-    let (mut styled, _) =
-        tokenize_with_grammar_skeleton(code, compiled, root_scope, light_theme, initial_stack)?;
-    if let Some(dark_theme_name) = html_theme.dark_theme_name.as_deref() {
-        if dark_theme_name == "none" {
-            styled = apply_dark_theme_inherit(styled);
-        } else {
-            let fallback_dark = default_theme_data(dark_theme_name);
-            let dark_theme = resolve_theme_data(dark_theme_name, themes).unwrap_or(&fallback_dark);
-            let dark_initial =
-                resolve_initial_stack(options_json, code, compiled, root_scope, dark_theme)?;
-            let (dark_styled, _) = tokenize_with_grammar_skeleton(
-                code,
-                compiled,
-                root_scope,
-                dark_theme,
-                dark_initial,
-            )?;
-            styled = apply_dark_theme_palette(styled, &dark_styled);
-        }
-    }
-    let default_fg = light_theme.fg.clone();
-    let mut lines = styled_json_lines(&styled);
-    normalize_vue_tag_end_tokens(&mut lines, root_scope, light_theme);
-    normalize_astro_tag_tokens(&mut lines, root_scope, light_theme);
-    apply_color_replacements_to_lines(&mut lines, &replacements);
-    let lines = apply_render_line_options(lines, &options)
-        .into_iter()
-        .map(|line| merge_line_for_html(&line, &default_fg))
-        .collect::<Vec<_>>();
-    Ok(render_styled_html_lines(&lines, &html_theme, false))
-}
-
-fn render_grammar_tokens_json(
-    code: &str,
-    options_json: &str,
-    compiled: &mut CompiledGrammar,
-    root_scope: Option<&str>,
-    themes: &HashMap<String, ThemeData>,
-) -> Result<String> {
-    let options: Value = serde_json::from_str(options_json).map_err(|err| {
-        Error::from_reason(format!("Failed to parse codeToTokens options JSON: {err}"))
-    })?;
-    let mut theme = resolve_theme_profile(options_json, "ferriki-grammar", themes);
-    let fallback = default_theme_data(&theme.theme_name);
-    let theme_data = resolve_theme_data(&theme.theme_name, themes).unwrap_or(&fallback);
-    let replacements = resolve_color_replacements(&options, &theme.theme_name);
-    apply_color_replacements_to_json_theme_profile(&mut theme, &replacements);
-    let initial_stack =
-        resolve_initial_stack(options_json, code, compiled, root_scope, theme_data)?;
-    let (styled, final_stack) =
-        tokenize_with_grammar_skeleton(code, compiled, root_scope, theme_data, initial_stack)?;
-    let mut lines = styled_json_lines(&styled);
-    normalize_vue_tag_end_tokens(&mut lines, root_scope, theme_data);
-    normalize_astro_tag_tokens(&mut lines, root_scope, theme_data);
-    apply_color_replacements_to_lines(&mut lines, &replacements);
-    render_styled_tokens_json_with_state(lines, theme, &final_stack, root_scope)
-}
-
-fn render_grammar_hast_json(
-    code: &str,
-    options_json: &str,
-    compiled: &mut CompiledGrammar,
-    root_scope: Option<&str>,
-    themes: &HashMap<String, ThemeData>,
-) -> Result<String> {
-    let options: Value = serde_json::from_str(options_json).map_err(|err| {
-        Error::from_reason(format!("Failed to parse codeToHast options JSON: {err}"))
-    })?;
-    let html_theme = resolve_html_theme_profile(options_json, "ferriki-grammar", themes);
-    let fallback_light = default_theme_data(&html_theme.theme_name);
-    let light_theme = resolve_theme_data(&html_theme.theme_name, themes).unwrap_or(&fallback_light);
-    let replacements = resolve_color_replacements(&options, &html_theme.theme_name);
-    let mut html_theme = html_theme;
-    apply_color_replacements_to_html_theme_profile(&mut html_theme, &replacements);
-    let initial_stack =
-        resolve_initial_stack(options_json, code, compiled, root_scope, light_theme)?;
-    let (mut styled, final_stack) =
-        tokenize_with_grammar_skeleton(code, compiled, root_scope, light_theme, initial_stack)?;
-    if let Some(dark_theme_name) = html_theme.dark_theme_name.as_deref() {
-        if dark_theme_name == "none" {
-            styled = apply_dark_theme_inherit(styled);
-        } else {
-            let fallback_dark = default_theme_data(dark_theme_name);
-            let dark_theme = resolve_theme_data(dark_theme_name, themes).unwrap_or(&fallback_dark);
-            let dark_initial =
-                resolve_initial_stack(options_json, code, compiled, root_scope, dark_theme)?;
-            let (dark_styled, _) = tokenize_with_grammar_skeleton(
-                code,
-                compiled,
-                root_scope,
-                dark_theme,
-                dark_initial,
-            )?;
-            styled = apply_dark_theme_palette(styled, &dark_styled);
-        }
-    }
-    let mut lines = styled_json_lines(&styled);
-    normalize_vue_tag_end_tokens(&mut lines, root_scope, light_theme);
-    normalize_astro_tag_tokens(&mut lines, root_scope, light_theme);
-    apply_color_replacements_to_lines(&mut lines, &replacements);
-    let lines = apply_render_line_options(lines, &options);
-    let rust_state = serialize_state_frames(&final_stack, root_scope);
-    render_styled_hast_payload_json(&lines, options_json, &html_theme, Some(rust_state))
 }
 
 #[napi]
@@ -5516,10 +379,15 @@ pub fn create_highlighter(options_json: String) -> FerrikiHighlighter {
 #[cfg(test)]
 mod tests {
     use super::*;
+    use crate::injection::*;
+    use crate::rule::*;
+    use crate::theme::*;
     use ferriki_asset_gen::{generate_catalogs_from_upstream, AssetSourceRef};
     use ferroni::scanner::{OnigString, Scanner, ScannerFindOptions};
+    use serde_json::json;
     use std::fs;
     use std::path::Path;
+    use std::sync::Arc;
     use std::time::{SystemTime, UNIX_EPOCH};
 
     fn temp_output_dir(label: &str) -> std::path::PathBuf {
